uhci.c didn't work well with USB storage. It would tend to stall relatively quickly and sometimes locked up the system. It usually only took me a couple of tries ripping a CD to reproduce the problem.
I took a long hard look at the locking in uhci.c and decided to clean
it up, fixing a couple of bugs along the way as well as documenting the
locking strategy.

With this patch applies, where I could only rip a CD a couple of times
before causing problems, I was able to rip a CD 12,000 times in a row
successfully, before I stopped it. Not a single error :)