[PATCH] Trivial APM patch: set_system_power_state
Number 5 from Thomas Hood

This patch renames the static function "apm_set_power_state"
to 'set_system_power_state'.

Generally, the prefix 'apm_' is required to prevent external
name collisions on exported functions.  This is a static function,
so the prefix isn't required for that purpose.  The prefix might
also indicate that this function has something particularly
to do with the apm subsystem; but that's not the case here.  This
function is simply a wrapper for set_power_state(), inserting the
argument which sets the power state for the whole system.
My main motivation for wanting to change this name is clearly
to indicate the difference between this function and
set_power_state().  Also, I would like to export set_power_state()
someday in the future, but this is a separate issue.