Handle $(export-objs) ambiguity
We use the makefile variable $(foo-objs) to list the objects
a composed module foo.o is supposed to be composed of.

We use the special varible $(export-objs) to list the object files which
export symbols.

This oviously clashes in the case of foo == export. There's basically
two ways to handle it: (1) rename one of these options, like 
foo-objs to foo-parts or something, or (2) simply disallow a composite
object called export.o, so you never need $(export-objs) to list its
parts.

As (1) would affect basically all Makefiles in the tree and (2) doesn't
seem much of a limitation, I went for (2).