[PATCH] cpia_usb, remove urb->next
This is the 2.5 version of the sample driver patch I sent around
last week.  It updates one video driver to work with the updated
ISO API:  no urb->next (explicit resubmit, with error detection)
and explicit transfer interval (HCs must remove 1 msec limit).

The 2.4 version is known to work, but usb video in 2.5 seems to
be a bit ill so this just fixes the compilation problem created
by the first of these patches.