[PATCH] set_cpus_allowed optimization
This adds an optimization to set_cpus_allowed: if the task is not
running, there is no sense in kicking the migration_threads into action,
we just need to update task->cpu.  This was suggested by Mike Kravetz.

Besides being an optimization, this would prevent any future race
between set_cpus_allowed and the migration_threads.