net/llc/*.c
Forward port of LLC from 2.4 to 2.5. This is the forward port of the LLC stack
released by Procom Inc. for Linux 2.0.30, I have heavily modified it to make
it similar to other Linux network stacks, using of struct sk_buff to represent
in-transit packets and doing massive code cleanups.

Jay Schullist contributed support for BSD Sockets, as the original code had
only a simple in kernel API for use by upper layer protocols, such as the
NetBEUI stack also provided by Procom for 2.0.30.

This code is basically what I had previously submitted to Alan Cox for his
2.4-ac series and that is even shipped in source form, in the Red Hat 7.3
kernel package, plus cleanups wrt standard syntax for labeled elements and
further use of this C construct to make the code more resilient to editing
mistakes, using the compiler to further check the source code.

TODO:

Make it completely SMP safe, as the reports of successful usage up to now and
the testing is done on UP.

Completely remove the old LLC code in the kernel, that is still there for things
like Appletalk, IPX, etc to use, also check that all these protocols work
correctly with this new LLC stack.

This code is already being used in the linux-sna project and Jay Schullist
has been developing support for things like DLSw and other protocols that works
on top of 802.2.

I'll be releasing patches with the NetBEUI stack and updated samba-2.0.6 patches
for use with NetBEUI and this LLC stack in the future. But the NetBEUI code
is available already in my kernel.org ftp area at:

ftp://ftp.kernel.org/pub/linux/kernel/people/acme.

Please report problems to me or the linux-sna mailing list, instructions on how
to subscribe are available at http://www.linux-sna.org website.