kbuild: Pass <target> to fixdep
cpp -MD guesses the target name from the source file name, but not
always correctly, e.g. when compiling fixdep, it generates prequisites
for fixdep.o instead of fixdep.
 
fixdep now fixes up the -MD generated output to contain the right target,
which is the one we pass on the command line.