ISDN: Eicon fix macro clash
The eicon header defined macros for byte, word, dword. Apart from that
not being a good idea in general, it did clash with the use of a parameter
called "word" in asm/signal.h. At some other place in the Eicon headers, 
there are also typedefs for these types, so the macros can simply go away.