kbuild: Introduce $(obj), $(src)
For separate source and object directories, Rules.make needs
to know where the files live.

For all the normal variables, $(obj-[ymn]) etc, it can figure that
out by itself, but for explicit rules to generate files it needs
help. 

So there $(obj)/target.o indicates that this file lives in the 
object directory, as opposed to $(src)/target.c, which lives in
the source dir.

For now $(obj) = $(src) = ., but convert some Makefiles to it
already.