kbuild: Add $(obj), $(src) for generated files
I suppose one can argue whether that is ugly or rather nice for
documentary purposes, but make on its own cannot figure out where a file
is supposed to live, so we have to help it.

(For the targets handled by Rules.make the situation is different, there
 Rules.make knows what's source and what's object and can add prefixes
 as necessary, so the most of the Makefiles are actually not affected by
 this kind of change)

For now, as $(obj) = $(src) = ., we only add "./", so the potential for
breakage is rather small.