[PATCH] IDE-101
Here is a quick fix.  I would like to synchronize with the irq handler
changes as well.  Becouse right now I know that preemption is killing
the disk subsystem when moving data between disks using different
request queues...  In esp.  It get's me in to do_request() with a queue
in unplugged state.  (Not everything is my fault, after all :-).