[PATCH] docbook: scripts/docproc improved [5/9]
This is the first patch in a serie to clean-up the DocBook
Makefile.

docproc is extented to include the functionality previously provided by
gen-all-syms and docgen.  Furthermore the necessity to specify which
files to search for EXPORT_SYMBOL are removed, the information is now
read in the .tmpl files.

docproc is furthermore extended to generate dependency information.
gen-all-syms and docgen are deleted.