[PATCH] usb_string fix
Things are indeed as conjectured, and I can reproduce the situation
where usb_string() returns -EPIPE. Now that this is an internal
error code for the USB subsystem, and not meant to get out to the
user, I made these driverfs files empty in case of error.
(While if there is no error but the string has length 0,
the file will consist of a single '\n'.)

One fewer random memory corruption. Unfortunately, there are more.

Andries