[PATCH] USB root hub polling and suspend
Currently with 2.5, when I suspend and resume my powerbook, I find
that the USB subsystem no longer sees root hub events, i.e. it doesn't
notice when I plug in a new USB device (it doesn't notice when I
unplug a device either but of course the driver for the device sees
that it is no longer responding).

It turns out that what happens is that the root hub timer goes off
after the OHCI driver has done its suspend stuff.  The timer routine
sees that the HCD is not running at the moment and doesn't schedule
another timeout.  Hence the series of timeouts stops.