[PATCH] PATCH: Support for Buffalo 40GB USB hard disk
About 3 weeks ago I sent the following information to the person listed as
the usb-storage maintainer (mdharm-usb@one-eyed-alien.net).  I have not
heard back from this person, and for completeness I thought I'd send you
the information at this time as maintainer of the USB stack in general.

In short, the message describes the additional entry in unusual_devs.h which
is needed to get a 40GB Buffalo USB hard disk working under Linux.  The
drive has been functioning in the PC now for over 3 weeks with no obvious
sign of problems.

The patch deals with revision "0113" which corresponds to the revision
reported by the drive I have access to.  I've included a short comment above
above the unusual_devs definition explaining why the entry is needed.  Feel
free to omit this if you feel it's unnecessary bloat.