kbuild: remove duplicated dependencies
Greg Banks:

A number of dep_bools and dep_tristates have architecture constants
(e.g. CONFIG_X86) as dependencies.  Such symbols have the property
that their value is constant for any particular arch tree, and is
either "y" or "", neither of which have any effect as dependencies,
so the dependencies have no effect.  This patch removes some of them.