[PATCH] md - 2 of 3 - Fix bug in raid5 AGAIN
That recent bug fix in raid5 just changed the bug, it didn't fix it.
I think that the original code was actually wrong, which didn't
help.

This time, the code actually matches the nearby comment, that has been expanded
a bit, so I feel somewhat more confident that it is actually right.