[LLC] change semantics of recvmsg and sendmsg
Now recvmsg will not discard the packet after reading less than
skb->len, keeping the rest of the skb data in receive_queue.

sendmsg will not barf if a packet bigger than the mtu is sent,
it'll politely only copy what is possible and tell this to
userland.

With this my openssh with PF_LLC works flawlessly 8) Time to
securely connect to a machine that has no ip address :-) This is
just to test how robust is PF_LLC, but sekurity guys may like it
for things like stealth IDS or bridge boxes with no IP addresses.