[PATCH] fix ext3 in data=writeback mode
When I converted ext3 to use to use direct-to-BIO writeback for
data=writeback mode I forgot that we need to hold a transaction open on
behalf of MAP_SHARED pages.  The fileystem is BUGging in get_block()
because there is no transaction open.

So let's forget that idea for now and send data=writeback mode back to
ext3_writepage.