[PATCH] sg might_sleep fixes
This is a update to a previous patch that fixed some sg might sleep
errors.

This patch corrects a problem in sg.c where a lock is held during calls to
vmalloc and calls for device model registration.

Note:
	Douglas Gilbert is the maintainer of this driver.
	dougg@gear.torque.net
	http://www.torque.net/sg/

	During Douglas Gilbert's time-off he connects when he can so it
	maybe a bit until he can address this.

	In the interim this patch should fix the problem, and
	still provide for safe additions.


The full patch is available at:
http://www-124.ibm.com/storageio/patches/2.5/sg

-andmike
--
Michael Anderson
andmike@us.ibm.com

 sg.c |   83 +++++++++++++++++++++++++++++++++++++++++--------------------------
 1 files changed, 52 insertions(+), 31 deletions(-)