[PATCH] remove get_free_page()
This used to be a macro wrapper around get_zeroed_page().

It was a dreadful choice of name, because it hid the fact that this
function memsets the page.  Probably, lots of the callers didn't
actually want to do that but this patch does not make such
optimisations.

At least now, people may spot a `get_zeroed_page' in there and realise
that the page doesn't actually need to be initialised.

The patch also changes get_zeroed_page() to go BUG if called with
__GFP_HIGHMEM.  It returns the address of the page's memory and hence
can only be used for direct-mapped pages (wli).