[PATCH] ips driver 5/6
2 minor bug fixes.

The first section makes sure we limit the size of the
sense_buffer copy to the target buffer's size so that
we don't overflow the sence_buffer.

The other sections remove some pointer arithmatic that
is wrong on 64bit machines do to padding.  Instead, just
call the pci_map functions on the buffer.