[PATCH] kNFSd: pre-zero response for lockd _msg requests.
All the lockd *_msg requests are implemented by
calling the equivalent non- *_msg request and sending off
the response as a new message (*_res).

For the non- *_msg requests, the response buffer is
pre-zeroed by the RPC layer, but the *_msg handlers weren't
pre-zeroing the buffer, would could in some circumstances lead to
an Oops.

Now we pre-zero them all.