[PATCH] usbcore doc + minor fixes
Cleaning out my queue of most minor patches:

  - Provides some kerneldoc for 'struct usb_interface' now that
    the API is highlighting it.

  - Fixes usb_set_interface() so it doesn't affect other interfaces.

    This provides the right place for an eventual HCD call to clean
    out now-invalid records of endpoint state, and also gets rid of
    a potential SMP issue where drivers on different interfaces
    calling concurrently could clobber each other.  (Per-interface
    data doesn't need locking except against config changes.)

  - It's OK to pass URB_NO_INTERRUPT hints if you're queueing a
    bunch of interrupt transfers.

The set_interface call should eventually take the interface as a
parameter, it's one of the few left using the "device plus magic
number" identifier.  I have a partial patch for that, but it doesn't
handle the (newish) ALSA usb audio driver or a few other callers.