[PATCH] 2.5.42 HID-BP menu
HID-BP confuses people, we know that.
I don't want to see the HID-BP drivers removed,
but I think it would be a good idea to make them more
difficult to enable, unless someone knows what they
are looking for.  (Basically it would reduce support
incidents a lot.)

The second way puts HID-BP in its own sub-menu and
updates the Help file.  I think that this is the better
choice, so unless there is a great hue and cry,
please apply the "hidbpmenu-2542.patch" file.