[PATCH] swim_iop
	* switched to private queue
	* set ->queue and ->private_data
	* switched to new methods
	* switched to use of ->bd_disk
	* fixed the idiocy I'd introduced there during end_request()
	  reshuffling - too many places got hit by search-and-replace ;-/

NOTE: swim_iop (and swim3) are broken for situations when several drives
are present.  Failure mode is rather amusing - the thing does a sort of
load-balancing and if you have two drives even requests end up sent to
the first and odd - to the second ;-) Had been there since they went
into the tree.  Swim3 mentions that support of multiple drives needs to
be fixed, swim_iop (derived from it) had lost even that warning.