[PATCH] 2.5.45 UCLinux merge M680x0 mmuless arch and include/asm
This is the M68K mmuless port code maintained by Greg Ungerer and used
by Snapgear for all sorts of bits of hardware. No shared files included
in this chunk of patches. This matches 2.5.45-uc1 which is the stuff that
has been in my tree for a bit combined with Kconfig changes