[PATCH] improved space efficiency in dcache
Currently we are storing filenames which are 16-chars or less
inside struct dentry itself and then separately allocating
larger names.

But this leaves spare space in the dentry - the dentry slab cache
is using cacheline alignment.   In my build, struct dentry is 112
bytes so there are at least an additional 16 bytes in there.

And the number of files which have names in the 16-32 char range
will be significant.

So Manfred's patch changes the dcache code to utilise _all_ the space
between the last member of the dentry and the start of the next cacheline.