[PATCH] get rid of ->init in osst
Since osst is the last driver still implementing ->init and Willem said
he's gonna do a major update including a resync with st anyway I think
it's okay to put this hack in for now.  Instead of ->init beeing
directly called from the midlayer osst_attach now calls in in the
beginning - it has an internal protection so that the initialization
will be called only one anyway.