[PATCH] Broken Hugetlbpage support in 2.5.46
The hugetlb page support in 2.5.46 is broken (Don't know if this is the
first version of kernel or any prior revs also have that).  Basically the
free side of hugepages was really not freeing the physical resources (for
the cases when the pages allocated using system call interface).  Attached
is the patch that should resolve it. (doesn't break the hugetlbfs support
either).