[PATCH] fix sd device number handling
This is the final patch to make sd work properly in a hotplug
enviroment.  Add a bitmap for currently used disks so that we
can properly reuse dev_t when a disk is hot-unplugged.

While at that add support for the eight additional majors
allocated to sd in the latest devices.txt [1].

If only register_blkdev finally died we could nuke almost all
knowledge of majors/minors from sd...

[1] http://www.lanana.org/docs/device-list/devices.txt