Deprecate /proc/pci
As discussed on lkml, /proc/pci is past his functional usefulness, and has
been obviated by lspci(8). lspci can provide all the same information, plus
so much more. And, it doesn't rely on having an in-kernel database of PCI
names. 

This change does: 
- Create config parameter: PCI_LEGACY_PROC.
- Add deprecation notice to Kconfig text.
- Wrap code in #ifdef in drivers/pci/proc.c

This does nothing with the names database, though we would probably be 
better off removing that as well. The PCI name field is used now only so 
drivers can print a useful message on startup - hardly a justification for
60kB + increase in size..