[PATCH] knfsd: Don't assume a reader on an RPC cache channel at statup.
Each cache channel remembers if there was a reader recently.
If there wasn't, then it doesn't bother with an up-call.
Currently, it initialises thinking there was a reader
recently incase the reader starts shortly afterwards.  However
this causes problems when a reader doesn't start soon, and
isn't necessary as the reader can easily be started before
the cache is initialised (e.g. before nfsd is started).