[PATCH] USB storage: remove usb_stor_tranfer_length()
This patch removes the (often troublesome) usb_stor_transfer_length()
function.

We've finally gotten all the command initiators to send the correct values
in the srb->request_bufflen field, so this is no longer needed.

There are probably some sanity checks that can also be removed now, but
that's for a later patch.