[PATCH] scanner.c, scanner.h: Remove PV8630 ioctls
This patch removes the inofficial ioctls that were used to support the
PV8630 USB-over-Parport chipset. They were alreaded ifdefed out.
Instead of them, the more generic (and official) SCANNER_IOCTL_CTRLMSG
should be used. The last software that used the old ioctl
(sane-hp4200) switched to the new ioctls a long time ago.

This patch is ontop of the "user-supplied" patch.