sysfs: make sure we drop all the references on dentrys we acquire.
None of the file deletions were dropping the referencees to dentrys that 
were acquired via the internal helper get_dentry(). This adds the missing
dput()s (and comments) to fix that.

Also, this adds a dget() on the directory and each file that's being removed
in sysfs_remove_dir(), (and associated dput()s) to guarantee the references
stay valid over the duration. 

Originally reported by Mika Penttilä.