kbuild: fix broken kallsyms on non-x86 archs
From: James Bottomley <James.Bottomley@steeleye.com>
  
kallsyms is broken in parisc on 2.5.56 again because of assembler syntax
subtleties.  This is the offending line:
  
printf("\t.byte 0x%02x ; .asciz\t\"%s\"\n"
  
Note the `;' separating the two statements.  On some platforms `;' is a
comment in assembly code, and thus the following .asciz is ignored.