[PATCH] USB ohci-hcd, don't force SLAB_ATOMIC allocations
This is a minor cleanup to let per-request memory allocations block,
when the caller allows (it provided the bitmask).  The driver used
to work that way until something like 2.4.3; an update (a few months
back) to how the "dma_addr_t" hashes to a "struct ohci_td *" lets us
simplify things again.  Another benfit:  it blocks irqs for less time
on the submit path.  (The ehci driver already acts this way.)