[PATCH] patch to DAC960 driver for error retry
Patch from Dave Olien <dmo@osdl.org>

The following patch implements retry on media errors for the DAC960 driver.
On such media errors, the DAC960 apparently doesn't report how much of the
transfer may have been successful before the error was encountered.

This type of error should be rare on healthy hardware, especially if the
disks are stripped or mirrored.  But, when large transfers are submitted to
the controller, it's especially bad to have to fail the entire transfer
because one disk sector may have been bad.