[PATCH] ext3: Remove journal_try_start()
journal_try_start() is a function which nonblockingly attempts to open a JBD
transaction handle.  It was added a long time ago when there were concerns
that ext3_writepage() could block kswapd for too long.

It was never clearly necessary.

So the patch throws it all away and just calls the blocking journal_start()
from ext3_writepage().