acpi: fix recently introduced proc-related bugs.
Recently, the acpi proc files were changed to use the seq_file interface.
As a result of this, the write methods became top-level write() methods
(i.e. not called by the proc layer). 

Whoever made these changes, did not take into account that the parameters 
passed to these changed. The fourth parameter previously was a void*, that
pointed to the private data for the owner of the file, and some of the 
methods still expected this, even though the parameter is now an loff_t*,
and points to the offset into the file that is being written. Huge Bug. 

This fixes the broken methods to look in the same place as the other 
methods (i.e. in the file's private data).