acpi sleep: divorce sleep functionality from power off functionality.
When ACPI turns the system off on shutdown, it actually enters S5, a sleep
state. This functionality is dependent on CONFIG_ACPI_SLEEP, which is 
dependent on CONFIG_SOFTWARE_SUSPEND.

This patch breaks the power off functionality into a separate file, and 
removes the dependency on the above-mentioned crap. Finally, power off works
for me again. 

Thanks to Tobias Ringstrom for the original patch.