[PATCH] sysfs error handling fix
Patch from: Rusty Lynch <rusty@linux.co.intel.com>

Attempting to cat a sysfs file that returns an error will result in an
endless dump of garbage to the screen because the result of the specific show
operation was being saved to a size_t (unsigned) and then later checked for a
negative value.

Here is a trivial patch to fix the error.