[PATCH] Make the world safe for -Wundef
Patch from: Valdis.Kletnieks@vt.edu

This is a patch to clean things up so compiling with -Wundef becomes feasible
(This patch cuts the number of warnings with 'make allyesconfig' drop from
around 10,000 to several hundred).  It was originally inspired by the
discussion of things that include linux/version.h but don't use the contents
- after doing this, I was able to find that there *WAS* at least one place
where version.h was missing (see following patch).