kbuild: Fix a race with module postprocessing
For the modversions case, we need vmlinux to be built before postprocessing
modules can start, since we need to extract the checksums from it.

For the non-modversions case, we use vmlinux if it's available (so we can
sensibly give warnings about unresolved symbols), but otherwise not. 
However, this could race with vmlinux being built at the exact same time as 
being processed.

Fixed by waiting until vmlinux is finished before starting module
postprocessing.