[PATCH] ext3: speed up O_SYNC writes
This is a forward-port of a 2.4 change from Stephen.  The (old) 2.5 code is
forcing a commit on every write by artificially dirtying the inode.

But generic_file_aio_write() has called generic_osync_inode() for us, which
has synced the file data.  There is no need to force the extra commit.