[PATCH] Move pfn_to_nid inline
Patch from William Lee Irwin

Inline and simplify pfn_to_nid - this is called heavily, it's a tiny
function, and makes a noticable difference in system time for kernel
compiles (sorry, lost the data). Is only used on NUMA machines.

Has been tested in my tree for over a month on UP, SMP, and NUMA and
compile tested against a variety of different configs.