[PATCH] hugetlb: fix MAP_FIXED handling
Patch from Rohit Seth

ia64 reserves virtual address region 4 (any address which starts with 0x4)
for huge pages.

Apparently, for hardware reasons, we do not wish to allow mappings with
other page sizes to appear in that region.

This patch prevents the user from being able to place regular MAP_FIXED
mappings into region 4 on ia64.  It is a no-op for ia32.