[PATCH] kbuild: Smart notation for non-verbose output
Create a nice shorthand to enable the non-verbose output mode.
make V=1        => Gives verbose output (default)
make V=0        => Gives non-verbose output

One of the reasons why people does not use KBUILD_VERBOSE=0 that
much is simply the typing needed.
This notation should make it acceptable to type it.
The usage of "make V=0" is restricted to the command line.
Anyone that wants to enable the non-verbose mode pr. default shall
set KBUILD_VERBOSE in the shell.