[ARM] NWFPE 2: Take advantage of the CPDO functions behaviour.
SingleCPDO, DoubleCPDO and ExtendedCPDO always set the destination
register type to their operating type (single, double, extended
respectively) when they complete successfully (nRc != 0).  This
means that we know what size result will be generated, and we don't
have to fetch it from the fType array.

In fact, since we know "nType" will be the type of the operand,
we don't even need to calculate value of Fd unless we need to
convert the result to the destination type.