[PATCH] revert "noirqbalance still doesn't do anything"
This recent patch caused Nick Piggin's 2xPIII VIA686B chipset machine into an
interrupts-off lockup during IDE probing.

We don't really know why - it might be because an interrupt is delivered to a
secondary which doesn't expect it.

I have a second patch from Zwane which solves the same problem in a different
way, but until that's had some wider testing I suggest we just back off the
original.