kbuild: build-targets replaced with always
The former name "kbuild-targets" did not give the user the intuitive feeling
that this target will be build regards less of any dependencies
no matter if we compile modules or not.
The new name "always" imply that this will be build always, and
tus is more informative.
Fixed the few users of build-targets, and added a warning in Makefile.build