Silently Ignore if the device is already active/disabled
Some drivers will try to activate a device even though it is already
active.  Instead of returning an error, the resource manager will now
just ignore this.  This should solve some of the recently seen problems.
Also it doesn't make sense to return an error if the device is already
in the correct state.