[PATCH] proc_file_read fix
From: Miklos.Szeredi@eth.ericsson.se (Miklos Szeredi)

This fixes a problem with method 0 of proc_file_read (when the whole file
is copied to the page).  The calculation of the final bytecount is wrong,
and hence smaller then page size reads will give a truncated file.

Current 2.4 kernels do it this way as well.