[PATCH] USB "Gadget Zero" driver (3/6)
This patch adds "Gadget Zero" (drivers/usb/gadget/zero.c).

Gadget Zero is a simple gadget driver that's useful for
testing controller drivers, and as an example to be used
for clone/modify style development.

This driver implements two configurations, and needs only
two bulk endpoints (in addition to ep0) ... so pretty much
any USB device controller should be usable with it in
one configuration or another.  It (optionally) supports
high speed devices, and has passed the USB-IF "chapter 9"
device model conformance tests.

It's worth noticing the kinds of hardware differences that
gadget drivers need to cope with.  Endpoints differ, in
ways that must be reflected various ways in descriptors.
And sometimes chip errata cause interoperability problems;
for example, an sa1100 can't change configurations after
enumerating.