[PATCH] ia32 subarch circular dependency fix
From: john stultz <johnstul@us.ibm.com>

This patch fixes a circular dependency (a function in mach_apic.h requires
hard_smp_processor_id() and hard_smp_processor_id() requires macros from
mach_apic.h) that has been in the subarch code for a bit, but was hacked
around with some #ifdefs.

With the inclusion of the generic-subarch the hack was dropped and bigsmp
and summit promptly broke.  So this makes things compile again.