[PATCH] USB: storage: abort and disconnect handling.
This patch re-organizes abort handling and enhances disconnect handling.

Not only do we keep track of the state (ABORTING, IDLE, etc.), but during
an abort we now introduce the idea of 'okay to send' or not.  The idea is
that we can now implement reset-after-abort properly.

We also track if we're disconnecting, and use that data to determine if we
can submit URBs or not.  Which means we can now disconnect during an abort.

This patch is from Alan Stern.