[PATCH] be more flexible about creating library archives
New makefile variable introduced: lib-y

The lib-y syntax allows you to do the usual tricks such as:

	lib-$(CONFIG_SMP) += percpu_counter.o

A built-in.o is always present in a directory that list .o files in
either obj-* or lib-*.

In contrast, lib.a is made only when lib-y is defined.

I also updated lib/Makefile, so that crc32.o is now always built-in
if selected.