[PATCH] USB: ehci, fix qh re-activation problem
This resolves a problem that appears when relinking
a bulk or control QH that has a partially completed
multi-packet qTD.  Some I/O could be repeated.

Such cases can happen when an empty QH starts to unlink,
but gets re-activated (by queueing the multi-packet qTD)
before the HC saw the unlink.  It's rarely an issue with
control traffic (transfers are so small) or when bulk
queues are active (the QH won't empty).