[PATCH] USB storage: unusual_devs fixups
This patch implements US_PR_DEVICE and US_SC_DEVICE, which have the meaning
'use the device's value -- no override'.  This should make maintance
easier, and also allow for those few devices that change their descriptors
depending on what they are connected to.

This will also print a message to help us identify entries that can be
pruned.

Finally, it removes a couple of dead flags.