[PATCH] dm: dm-ioctl.c: Unregister with devfs before renaming the device
DM originally stored a devfs handle in the hash-cell, and performed the
unregister based on that handle. These devfs handles have since been removed,
and devices are registered and unregistered simply based on their names. So
the device now needs to be unregistered before we lose the name.

See the following BK change for more details:
http://linux.bkbits.net:8080/linux-2.5/diffs/drivers/md/dm-ioctl.c@1.6?nav=index.html|src/|src/drivers|src/drivers/md|hist/drivers/md/dm-ioctl.c
[Kevin Corry]