[PATCH] sd.c: set data direction to SCSI_DATA_NONE for START_STOP
while trying to access a disk drive via an FCP bridge we got an FCP_RSP IU
with the RSP_CODE field set to "FCP_CMND Fields Invalid". This happened
after sending a START_STOP command to the device. Reason for this was that
the FCP_CMND IU incorrectly had the RDDATA field set to one, because of a
bug in sd_spinup_disk(). There the data direction for START_STOP is set
to SCSI_DATA_READ instead of SCSI_DATA_NONE.
Please apply the patch below.

Thanks,
Heiko