o net: make sk_{add,del}_node functions take care of sock refcounting
With this we make it easier to write correct network families as less
details need to be taken into account, as well in the current state we
make the non-refcounting protocols (the ones still keeping deliver_to_old_ones
in the tree) suck less. 8)

Left a WARN_ON in sk_del_node_init for a while, so that we can catch cases
where we're using __sock_put on a struct sock that has refcnt == 1, which
is not the case for all the ones I tested.