[PATCH] Remove spinlock workaround for pre 2.95 gccs
Remove the empty initializer workaround that was added for egcs 1.1.
Only 2.95+ is supported now, so all compilers should support empty
structures.

The if just checked for __GNUC__, which means that 2.95 got
the workaround (and the incompatibility) too even though it didn't need it.

Advantage is that gcc 2.95 and 3.x compiled kernels are now potentially
binary compatible.  Module loading still checks the compiler version,
but it might be removable.