[PATCH] USB: make kaweth deal with ENOMEM
this fixes the logic kaweth uses to deal with ENOMEM from
usb_submit_urb(). Using the interrupt endpoint's completion
handller is longer an option because automagic resubmission
has been removed. Thus we use workqueues which can be delayed
so that kswapd can do its job and can use GFP_NOIO which is much
likelier to succeed.