[PATCH] USB: usbnet, don't NET_XMIT_DROP
Martin Diehl pointed out that the network layer treats
a tx drop notification in special ways, so "usbnet"
shouldn't be using them.  This fixes that problem,
which could potentially explain some infrequent oopsing
seen in the network layer.  (With memory so tight that
an URB can't be gotten.)

Please merge.