[PATCH] USB: ftdi_sio - additional pids
Wayne Wylupski sent some more PIDs (in a patch, but I've modified
it) for the ftdi_sio driver to the ftdi-usb-sio-devel list.  These
PIDs were for additional CrystalFontz LCD displays.

An earlier patch from David Glance for the DSS-20 SyncStation was
applied to the 2.6 tree.  This did not have a changelog entry in
the file header comment of ftdi_sio.c, so I have retroactively
added one in the attached patch.