[PATCH] SCSI dc395x compile problem
I'm besides 2 others one of the Maintainers of the SCSI DC395x SCSI
driver and I found a little 1 char problem during the merge from the AC
tree to the official BK tree which causes a compile issue.

Here is a little fix for the 2.6.0-test2-bk1 that allows compiling
correctly. I would be thankful if this patch can be included in an
upcoming bk2 or test3 release.