[PATCH] zap_other_threads() detaches thread group leader
The change to detach the threads in zap_other_threads() broke the case
where the non-thread-group-leader is the cause of de_thread().

In this case the group leader will be detached and freed before
switch_exec_pids() is complete and invalid data will be used.  This is a
patch that makes sure that the group leader does not get detached and
reaped.