Add new pmac_zilog serial driver, obsolete old macserial.
The new driver is a complete rewrite based on David Miller sunzilog
adapted to PowerMac, it uses the new driver model & the serial driver
core unlike the old macserial. It doesn't support DMA yet but this is
a "feature" for now as the DMA implementation of macserial used to
exhibit memory corruption problems.