[PATCH] check whether a disk got writeable in sd_open
This is the 2.5 version of a 2.4 patch posted to the list long
ago, the aacraid thread reminded me of it.

The problem is that certain highend arrays allow to mark a r/o
volume writeable on the fly so we have to call check_disk_change
for write-protected devices in sd_open, too.