[PATCH] kbuild: warn if the user has old modutils
From: Valdis.Kletnieks@vt.edu, Sam Ravnborg <sam@ravnborg.org>

Adds an explicit check for the new modutils in the build system.

Generally we should avoid these sorts of hardwired checks for the right
versions of things, but we are still getting a significant number of problem
reports due to people not having the new tools.  Let's help them out.