[PATCH] zoran driver documentation fix
From: Ronald Bultje <rbultje@ronald.bitfreak.net>

the zoran kernel driver is called 'zoran.o' in its CVS (historical thing, I
don't know why), and it's called zr36067.o in the kernel tree.  The
documentation in the kernel tree refers to zoran.o, though, which is (in
the kernel tree) the driver for zr36120-based cards, rather than the driver
for zr360x7-based cards.

The attached patch fixes the documentation and makes it refer to
zr36067.o instead.