[PATCH] make CONFIG_HUGETLB_PAGE mirror CONFIG_HUGETLBFS
At present you can set CONFIG_HUGETLB_PAGE and not CONFIG_HUGETLBFS, which
give the kernel low-level hugepage support, but no userspace API with which
to access it.

And with the recent hugetlbfs-accounting fix, the low-level code needs
functions which are in hugetlbfs, so CONFIG_HUGETLBFS=n,
CONFIG_HUGETLB_PAGE=y doesn't even link.

So we flip things around: CONFIG_HUGETLB_PAGE is set if CONFIG_HUGETLBFS is
set and CONFIG_HUGETLB_PAGE is unset if CONFIG_HUGETLBFS is unset.  The
CONFIG_HUGETLB_PAGE option hence disappears from the per-arch configuration
menus.