[PATCH] ftape linkage error
Since 2.6.0-test6, ftape can't be configured as a built-in driver.
test6 changed ftape-init.c to call ftape_proc_destroy() also in
the non-MODULE case; however, ftape_proc_destroy() is only defined
when the driver is built as a module. The result is a linkage error.

This fixes this by deleting the #if MODULE around ftape_proc_destroy()'s
definition.