[PATCH] Fix initrd with devfs enabled
This fixes initrd with devfs.  With that combination the late-boot code
does temporary mount of devfs over rootfs /dev, which made /dev/initrd
inaccessible.  For setups without devfs that didn't happen. 

The fix is trivial - put the file in question outside of /dev; IOW,
we simply replace "/dev/initrd" with "/initrd.image" in init/*.

Confirmed to fix the problem by Valdis Kletnieks