[PATCH] I2C: i2c documentation (1 of 2)
This is the document I wrote (and you reviewed) about porting client
drivers to Linux 2.6. The retained name is "porting-clients" (in line
with writing-clients). I won't commit it to i2c/lm_sensors2 CVS, since
that document is of no use outside of the 2.6 kernel (and I'm bored
keeping files in sync).