[ACPI] Update Linux to ACPICA 20031203 (Bob Moore)
Changed the initialization of Operation Regions during subsystem init to
perform two entire walks of the ACPI namespace; The first to initialize
the regions themselves, the second to execute the _REG methods.  This
fixed some interdependencies across _REG methods found on some machines.

Fixed a problem where a Store(Local0, Local1) could simply update the
object reference count, and not create a new copy of the object if the
Local1 is uninitialized.

Implemented support for the _SST reserved method during sleep
transitions.

Implemented support to clear the SLP_TYP and SLP_EN bits when waking up,
this is apparently required by some machines.

When sleeping, clear the wake status only if SleepState is not S5.

Fixed a problem in AcpiRsExtendedIrqResource() where an incorrect
pointer arithmetic advanced a string pointer too far.

Fixed a problem in AcpiTbGetTablePtr() where a garbage pointer could be
returned if the requested table has not been loaded.

Within the support for IRQ resources, restructured the handling of the
active and edge/level bits.

Fixed a few problems in AcpiPsxExecute() where memory could be leaked
under certain error conditions.

Improved error messages for the cases where the ACPI mode could not be
entered.