[PATCH] Fwd: Re: Atmel - possible SKB leak?
Jeff,

Atmel driver in 2.6.0-test11 is leaking SKBs if card gets disassociated
from an AP when it's about to transfer packet. Simon (atmel maintainer)
is OK with the patch. Given the fact that we are leaking memory I think
it may be beneficial to push it to Linus (if you like the patch).

Dmitry

===================================================================


ChangeSet@1.1517, 2003-12-11 01:44:56-05:00, dtor_core@ameritech.net
  NET: atmel - do not leak SKBs when dropping packets


 atmel.c |    6 ++++--
 1 files changed, 4 insertions(+), 2 deletions(-)


===================================================================