[PATCH] lib/inflate.c fix
From: "H. Peter Anvin" <hpa@zytor.com>

This patch fixes the "non-terminating inflate" problem that Russell King
complained about on LKML earlier today.

I chose to use "goto" much like zlib does, in order to not require
setjmp/longjmp inside the kernel.  It's a bit ugly, but it also lets each
function chose how it needs to be terminated on error, which is a good
thing.