[PATCH] remove dev_get from wanrouter
The call to dev_get() in wanrouter_device_new_if is racy and redundant and should
be removed.  The later 'register_netdev()' does the same test internally and will
return the appropriate error if the name already exists.

This patch is against 2.6.0-test6.
Resend of earlier patch because it was ignored, or missed.