[PATCH] remove old Eicon isdn driver
From: Armin <armin@melware.de>

The new driver for Eicon ISDN cards is in drivers/isdn/hardware/eicon.

The old driver is still in drivers/isdn/eicon and I would like to remove
this old driver from the 2.6 kernels.

The attached patch removes the references to this driver, but in addition
the directory drivers/isdn/eicon should be removed.