[PNP]: Card matching code fix
This patch updates the matching code to ensure that all requested
devices are present on the card, even if they are in use.  It is
necessary for some ALSA drivers to work properly because early vendors
would have different sets of devices on the same card ids.  It is from
Takashi Iwai <tiwai@suse.de>.