[ARM PATCH] 1759/1: Add ARM925 support, updated
Patch from Tony Lindgren

The following patch adds the ARM925 processor support.

This patch is based on the proc-arm925.S in 2.4 kernel, and it has been
modified to be more like the ARM926 support in the 2.6 kernel, except for
the cache flushing and initialization. Also, this patch fixes the
writeback cache flushing for most part.

There are still some issues with writeback cache flushing with devices 
using direct memory access, such as USB OHCI. It is recommended to use
the writethrough instead if using USB. This is the default for OMAP-1510.
OMAP-1610 does not have these issues, as it uses ARM926 core.

This patch includes the update for the processor functions to be armv4t 
instead armv5t as suggested here:

http://www.arm.linux.org.uk/developer/patches/viewpatch.php?id=1740/1