[PATCH] I2C: rename sysfs files, part 2 of 2
Here is the second step of my sysfs renaming plan.

This one does the following renames (as I already announced on the
LKML):
temp<n>_hyst  -> temp<n>_max_hyst or temp<n>_crit_hyst
sensor<n>     -> temp<n>_type
pwm<n>        -> fan<n>_pwm
pwm<n>_enable -> fan<n>_pwm_enable
vid           -> in<n>_ref

The associated libsensors patch is here:
http://jdelvare.net1.nerim.net/sensors/libsensors-sysfs-names-2.diff
(not applied yet, on purpose)


Note that the w83781d part is a bit more complex, not only because it is
the only driver to require the 5 changes, but also because at some point
the macros assume that the internal variable names match the sysfs
names, so I had to change them too (better than rewriting the macros,
methinks).

For reference, here is the list of changes, by driver:

asb100:
    hyst -> max_hyst (4)
    pwm -> fan_pwm (1)
    pwm_enable -> fan_pwm_enable (1)
    vid -> in_ref (1)
fscher:
     pwm -> fan_pwm (3)
gl518sm:
    hyst -> max_hyst (1)
it87:
    sensor -> temp_type (3)
lm75:
    hyst -> max_hyst (1)
lm78:
    hyst -> max_hyst (1)
    vid -> in_ref (1)
lm85:
    pwm -> fan_pwm (3)
    pwm_enable -> fan_pwm_enable (3)
    vid -> in_ref (1)
lm90:
    hyst -> crit_hyst (2)
via686a:
    hyst -> max_hyst (3)
w83781d:
    hyst -> max_hyst (2)
    sensor -> temp_type (3)
    pwm -> fan_pwm (4)
    pwm_enable -> fan_pwm_enable (1)
    vid -> in_ref (1)

There's also a documentation update. There will be one more after that
(to make it more readable, no contents change), and a patch to lm83.c to
bring it to compliance with the (new) standard. (If you wonder why I did
not change it with the other drivers: because it was *already* not in
compliance with the old standard. There's some real work to do for this
one.)