Make the SCSI mempool allocations variable
We still retain the default MAX_PHYS_SEGMENTS (which
is currently 128).  However, a mechanism for raising
the limit correctly is added for people who wish to
benchmark with larger sglists.

A decision on how we present this to the user will be
taken at a later time when more information about
the usefulness of a higher limit is available