[PATCH] move consistent_dma_mask to the generic device
From: James Bottomley <James.Bottomley@SteelEye.com>

pci_dev.consistent_dma_mask was introduced to get around problems in the
IA64 Altix machine.

Now, we have a use for it in x86: the aacraid needs coherent memory in a
31 bit address range (2GB).  Unfortunately, x86 is converted to the dma
model, so it can't see the pci_dev by the time coherent memory is
allocated.

The solution to all of this is to move pci_dev.consistent_dma_mask to
dev.coherent_dma_mask and make x86 use it in the dma_alloc_coherent()
calls.

This should allow me to make the aacraid set the coherent mask instead
of using it's current dma_mask juggling.