[PCMCIA] Add, fix, update PCMCIA debugging.
The PCMCIA core debugging assumes that "pc_debug" is a global.  If
pc_debug were to be a global, it would pollute the kernel's global
name space, potentially clashing with PCMCIA drivers and even ds.c.
 
We also get rid of PCMCIA using the "DEBUG" macro, which is the
standard macro to enable debugging in the kernel.  Instead, we
introduce cs_dbg() for core PCMCIA debugging, ds_dbg() for driver
services debugging, and similar macros in the socket drivers.