[PATCH] ppc32: Fix PCI DMA API changes
From: Tom Rini <trini@kernel.crashing.org>

When Dave Miller updated <asm-ppc/pci.h> for the _for_device and _for_cpu
portions of the PCI DMA API, he assumed that on PPC32 consistent_sync*
would also need to be changed for device or cpu.  This is not the case, so
what this does is the pci_*_for_{cpu,device} calls call
consistent_sync{,page} again.