[PATCH] ppc32: Fix booting some IBM PRePs
From: Tom Rini <trini@kernel.crashing.org>

The following patch comes from Paul Mackerras.  Earlier on in 2.6,
arch/ppc/boot/utils/mkprep.c was changed slightly so that it would build
and work on Solaris.  Doing this required changing from filling out
pointers to an area to filling out a local copy of the struct.  However, a
memcpy was left out, and the info is only needed on some machines to boot.
The following adds in the missing memcpy and allows for IBM PRePs to boot
from a raw floppy again.