[PATCH] Fix uninlined memcmp on i386
From: DHollenbeck <dick@softplc.com>

This patch was needed against a pristine 2.6.4 kernel when compiling with
"gcc 3.4 _very recent_" using the -Os option.

Without this patch, modules would use a non-inline memcmp() and then not
find it in the kernel, causing depmod to complain and some modules not to
load.