[PATCH] Hotplug CPUs: Set prio of migration thread before CPU
We need the migration thread to be RT as soon as the CPU comes online:
for example, stop_machine() (another RT task) expects to yield to it.
Extract the core of setscheduler() and do that when the migration
thread is created.  rq lock is a precaution against the (theoretical)
possibility of someone else doing setscheduer on this thread at the
same time.