[PATCH] sh: port sh-sci driver to the new API
From: Paul Mundt <lethal@linux-sh.org>

This ports the sh-sci driver to the new API.  sh and h8 both use this.  The
intention is to leave the drivers/char sh-sci in place for a short period of
time until all the h8 people are on the new driver (sh no longer uses the
drivers/char version).