[PATCH] dmasound close timeout
The dmasound driver occasionally hangs a process on exit.

Apparently there is a possible case where the sound HW stops draining
output samples and the driver waits forever in its release() callback.
It should check for signals(), but it seems signal_pending() never
returns 1 when the process is beeing killed (implicit release() of files
on exit).

This patch adds a safety timeout to the release() function to make sure
we can at least close the driver.  I'll try to find the reason we aren't
driving samples later, but it is better to have a safety just incase the
sound clock goes berserk for some reason.