[PATCH] ppc32: arch code preempt fixes
I figured the best way to stop beeing bothered by users trying to run
preempt was to fix it ;)

Here's a first batch that close some races we had when testing regs->msr
for altivec or FPU enable, then doing the giveup_* function.  A preempt
in between those would have caused us to save a stale altivec or FPU
context.