[PATCH] consolidate compat_sys_mount
This replaces six duplicated implementations of various quality of
sys32_mount with a shiny new compat_sys_mount().

It's been tested on parisc64 and sparc64 and fixes a bug exposed by the
latest revision of Debian's initscripts.  Thanks to Arnd Bergmann and
Dave Miller for their suggestions, fixes and testing.  Please apply.