[PATCH] siginfo.si_band is long
From: Marcus Meissner <meissner@suse.de>

After discussion on the glibc list the result was that=20 si_band is "long
int" according to POSIX:

	http://www.opengroup.org/onlinepubs/007904975/basedefs/signal.h.html

Ulrich Drepper refused a patch to fix glibc due to this reason:
	http://sources.redhat.com/ml/libc-alpha/2004-03/msg00254.html

so here is the patch to fix it in the kernel.

ppc64 and s390x were broken before and are fixed by this patch too.