[PATCH] Fix genksyms parsing
From: Rusty Russell <rusty@rustcorp.com.au>

From: Andreas Schwab <schwab@suse.de> I'm getting a warning when building
for ia64 with MODVERSIONS enabled.  This is a bug in genksyms, it can't
cope with some arguments of __typeof__.

The following patch will fix that.  Actually the argument of __typeof__ is
an abstract declarator, but the genksyms parser has no production for that;
decl_specifier_seq also matches some invalid constructs, but I don't think
this is a problem in practice, since the compiler will reject them.