[PATCH] USB: Implement endpoint_disable() for UHCI
This patch implements the endpoint_disable method for the UHCI driver, as
you requested a while back.  It guarantees that during unbinding events
(disconnect, configuration change, rmmod) the UHCI driver will have
finished using every URB for the interface being unbound.  It doesn't
quite guarantee that the completion handlers will have finished running,
but it would take a pretty unlikely race to violate that assumption.  (I
think it's the same with the OHCI and EHCI drivers.)

Despite the patch numbering this one applies _after_ as249, which is a
more important bugfix.