[libata] clean up taskfile submission to hardware
When writing taskfile (an ATA command) to the controller, the exact
setup of the taskfile is dependent on the taskfile "protocol":
PIO, PIO Multiple, DMA, Non-data, etc.

So, we separate out the submission of the taskfile to hardware into
a separate function ata_qc_issue_prot(), which will later be the
home for more code.

Also, remove some dead code (#if 0'd).