[PATCH] USB: USB altsetting updates for IDSN Hisax driver
The USB core is changing the way interfaces and altsettings are stored.
They are no longer required to be in numerical order, and as a result,
simply indexing the interface and altsetting arrays won't work as
expected.

This patch for the st5481 takes these changes into account.  A simpler
approach would be to store a pointer to the struct usb_host_interface
rather than look it up repeatedly, but I'm not very familiar with this
driver and didn't want to attempt such an alteration.