[PATCH] ppc32: pmac support update
This adds some initial support for the latest model of iBook G4 (still
need some work on the clock chip at least and some radeonfb updates that
I'll send later along with other fixes for this driver). 

It also removes a useless delay and fixes detection of the airport card
on the "Windtunnel" class desktop G4 machines.