[PATCH] simplify mqueue_inode_info->messages allocation
From: Chris Wright <chrisw@osdl.org>

Currently, if a user creates an mqueue and passes an mq_attr, the
info->messages will be created twice (and the extra one is properly freed).
This patch simply delays the allocation so that it only ever happens once. 
The relevant mq_attr data is passed to lower levels via the dentry->d_fsdata
fs private data.  This also helps isolate the areas we'd need to touch to do
rlimits on mqueues.