[netdrvr] remove rcpci driver, for Red Creek Hardware VPNs
Pete Popov, the author says

This driver is obsolete and broken in 2.4 (and I'm pretty sure in 2.6).
The hardware has not been available for a while. I wrote the driver
for 2.2 but after I left RedCreek (which went out of business),
someone in the community updated it to 2.4, but I don't think that
person even had the hardware to test it so the driver remained broken
in 2.4. So my recommendation is that it's really time to remove this
driver from the kernel tree. Just my opinion but I thought I'd share
it with you.
[...]
I can't imagine that there are any users because 2.4 was broken last
time I checked (admittedly that was a year ago).