[PCMCIA] Convert IO resource allocation to use struct resource.
This causes PCMCIA to use struct resource internally for IO resources.
This means that we can keep track of the resource pointer, expand
this resource if necessary, and use release_resource() on it when
we're done.

This eventually means that we can change these to normal resources
which aren't marked busy.