[PATCH] getblk() BUG removal
We keep on getting BUG()s from isofs_read_super() because it passes an insane
blocksize to bread().  See http://bugme.osdl.org/show_bug.cgi?id=2735 for
example.

I don't know what's up with isofs, but going BUG in there seems a bit rude.
Change it to drop a bunch of diagnostics and a backtrace then return a null
bh*.

Most callers of getblk() don't expect it to fail, so they'll oops anyway.  But
isofs does actually check for a NULL return.  This way, the machine stays up
and we get better debug diagnostics.