[PATCH] USB: usb retry cleanups
This patch stops changing the reported fault mode in cases where retries of
GET_DESCRIPTOR fail because the device just doesn't have such a descriptor.
Plus, it stops printing messages when retrying.

It also reduces the number of retries; the first retry seems to resolve most
of these firmware problems.

Signed-Off-By:  David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>