[PATCH] Make paca xCurrent field be a pointer
The paca struct contains a pointer to the current task, which is used
for the `current' macro.  For some reason, this field is a u64, and
every time we use it we need a cast, because it is really a pointer.
This patch cleans things up a little by making it a pointer to struct
task_struct and removing the casts.  It also removes a now-incorrect
comment which said that r13 contains current (it now holds
&paca[smp_processor_id()]).

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>