Fix x86 "clear_cpu()" macro.
We need to clear all exceptions before synchronizing
with the FPU, since we aren't ready to handle a FP
exception here and we're getting rid of all FP state.

Special thanks to Alexander Nyberg for reports and
testing. Alternate patches by Sergey Vlasov and Andi
Kleen, who both worked on this.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>