[SELINUX]: Fine-grained Netlink support - SELinux headers
This patch regenerates the SELinux module headers to reflect new class
and access vectors definitions.  The size of the diff is misleading;
much of it is simply a change in the ordering of the automatically
generated definitions. The corresponding generation script has been
changed to ensure a stable order in the future.

Author: Stephen Smalley <sds@epoch.ncsc.mil>
Signed-off-by:  Stephen Smalley <sds@epoch.ncsc.mil>
Signed-off-by:  James Morris <jmorris@redhat.com>
Signed-off-by:  David S. Miller <davem@redhat.com>