[PATCH] x86: remove APIC_LOCKUP_DEBUG
the patch below gets rid of APIC_LOCKUP_DEBUG.  It has been in the kernel
for more than 3 years and the message was only reported once during that
period of time - and even in that case it was a side-effect of a really bad
crash.  The lockup workaround works, the debugging code can be moved out.

Signed-off-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>