[PATCH] 3c59x: support for ATI Radeon 9100 IGP
The following patch adds support for the 3Com networking core found in the
ATI Radeon 9100 IGP southbridge used on boards like Asus P4R800-VM.

The main point of this patch is using HAS_MII instead of HAS_NWAY for the
definition of the board.  All the previous trials since the end of last
year used HAS_NWAY which disables the Tx part of the transceiver; using
HAS_NWAY was the way all 3Com Cyclone and Tornado chips worked, as they had
the transceiver integrated.  The ATI solution has an external transceiver
and I had to physically see the different chip on the board (the board was
provided by ATI) to finally understand that it needs the HAS_MII
definition...  I'm still waiting for some docs from ATI to clarify if this
is the correct way of handling this chip and if there are any differences
w.r.t EEPROM handling, but as it appears to work and was also confirmed by
other testers, I don't want to keep owners of such boards away from their
networks :-)

The textual identification was a bit hard to decide; it's called
"3c920B-EMB-WNM" in the Windows .INF file that Asus provides for their
boards.  As this name was already used for PCI ID 9210, I added the
paranthesis which specifies where this chip is found.

The Scyld driver defines FEATURE_TORNADO to include HAS_NWAY.  This board
would then probably need to not be defined with FEATURE_TORNADO, but the
same as in this patch.

I would like to publicly thank Tyson Vickers for both ideas and patience
during the last few weeks.  He managed to get the driver working by
randomly setting driver parameters :-) But then he contacted me and worked
with me towards the solution.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>