[PATCH] fix handling of '/' embedded in filenames in isofs
 * Fix slashes in broken Acorn ISO9660 images in fs/isofs/dir.c (Darren Salt)
This fixes Debian BTS #141660.
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=141660

	From: Darren Salt <linux@youmustbejoking.demon.co.uk>
	Message-ID: <4B238BA09A%linux@youmustbejoking.demon.co.uk>
	To: submit@bugs.debian.org
	Subject: Handle '/' in filenames in broken ISO9660 images

[Also applicable to 2.2.x]

There has been for some time a problem with certain CD-ROMs whose images
were generated using a particular tool on Acorn RISC OS.  The problem is
that in certain catalogue entries, the extension separator character '/'
(RISC OS uses '.' and '/' the other way round) was not replaced with '.';
thus Linux cannot properly parse this without this patch, thinking that it
is a directory separator.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>