[PATCH] USB: dummy_hcd shouldn't reject SET-ADDRESS requests
This patch for the dummy_hcd driver prevents it from rejecting SET-ADDRESS
requests when the address is already set.  The USB 2.0 spec states that if
a device is in the ADDRESS state, it should accept and start using the new
address.  Behavior in the CONFIGURED state is undefined, but since
dummy_hcd doesn't keep track of the difference between the two states we
might as well accept the new address in any case.


Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>