[PATCH] consolidate in-kernel configuration
From: Andy Whitcroft <apw@shadowen.org>

Being able to recover the configuration from a kernel is very useful and it
would be nice to default this option to Yes.  Currently, to have the config
available both from the image (using extract-ikconfig) and via /proc we
keep two copies of the original .config in the kernel.  One in plain text
and one gzip compressed.  This is not optimal.

This patch removes the plain text version of the configuration and updates
the extraction tools to locate and use the gzip'd version of the file.
This has the added bonus of providing us with the exact same results in
both cases, the original .config; including the comments.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>