[PATCH] ia64: trashing bootmem on non-NUMA boxes
call_pernode_memory() is rounding memory regions to page boundaries,
but in the non-NUMA case, it's incorrectly passing the rounded start and
the original length.  Any reserved memory regions sitting beyond a
memory region that isn't nicely aligned is potentially at risk for
getting freed and trashed.  This is a fairly nasty bug that can bite in
very hard to find ways.  In my case, the EFI memory map was getting
overwritten, causing the ACPI methods to access chipset registers
cachable instead of uncachable.  It could potentially be causing some of
the random hangs and strange behavior we're seeing lately.

Signed-off-by: Alex Williamson <alex.williamson@hp.com>
Signed-off-by: David Mosberger <davidm@hpl.hp.com>