[PATCH] PPC64 iSeries fails to boot
Jeremy Katz of Red Hat reported that his iSeries machine would not boot
with 2.6.7 based kernels.

It appears that with the inclusion of Paul Mackerras' patch "Optimize
exception/syscall entry/exit" a small previous patch got reverted.  Here
is that patch again.  The lack of this patch does not stop all iSeries
machines booting, but it does stop some.

Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>