[PATCH] ppc64: janitor log_rtas_error() call arguments
This patch from Linas Vepstas (rediffed by me) fixes the confusing argument
aliasing of the log_rtas_error() subroutine.

This patch makes no functional changes, it just cleans up some strange
usage.

The rtas_args used to communicate with firmware are always taken from the
paca struct, so as to keep the args at a fixed, low-memory location.  But
the log_rtas_error() routine also took an rtas_args pointer, which it
assumed was aliased to the paca struct.  This aliasing is both
un-neccessary, and confusing; this patch eliminates this confusion.

Signed-off-by: Linas Vepstas <linas@linas.org>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>