kbuild: Less intrusive LANG override, fixes menuconfig
The locale override caused problems for some people with locale setiings
different from 'C'. make menuconfig was looking bad / unuseable.
This patch limit the override of locales to the part where we actually descend the kernel
doing the full build of the kernel.
The speed improvement is the same.

make menuconfig should now be useable for all locale settings again.
Thanks to Marcel Sebek <sebek64@post.cz> for pointing out this problem and being
paitent in testing.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>