[NET]: Allow MD5 to be a module
I found that recent 2.6 kernels no longer allowed me to build MD5 as
a module even though everything that used it were modules (including
ipv6 and sctp).  It turns out that there were boolean options
selecting MD5 in the Kconfig files.  Due to limitations in the current
kconfig implementation, this forces MD5 to be a boolean as well.

The usual workaround in these cases is to move the selection up
to the closest tristate.  This is what the following patch does.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>