[PATCH] con_font_copy sanitized
->con_font_copy() sanitized.  We extract the number of console to copy the
font from in the caller (it's taken from the field of console_font_op that
is normally used for font height - messy even for an ioctl, but that animal
used to be passed all the way down into console drivers).

With decoding done in con_font_copy(), we simply pass source console number
into the method.

Signed-off-by: Al Viro <viro@parcelfarce.linux.org.uk>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>