[PATCH] USB: Make removable-LUN support a non-test option in the g_file_storage driver
This patch follows the suggestions sent by Todd Fischer and Diego Dompe
for making removable-LUN support part of the normal non-testing version of
the g_file_storage driver.  It also moves LUN device registration to the
correct place and eliminates a code path that stalls the bulk-out pipe in
a racy way.

There are also some smaller changes: update some comments, add initial
debugging support for USB suspend/resume, and miscellaneous code cleanups.
Last but not least, the driver has been sufficiently stable for
sufficiently long that it's fair to remove the "(DEVELOPMENT)" warning in
Kconfig.



Sent-by: Todd Fischer <toddf@cadenux.com>
Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>