[PATCH] Fix BSD accounting cross-platform compatibility
BSD accounting cross-platform compatibility is a new feature of 2.6.8 and
thus not crucial, but it'd be nice not to have kernels writing wrong file
formats out in the wild.

The endianness detection logic I wanted to suppose for userspace turned out
to be bogus.  So just do it the simple way and store endianness info
together with the version number.

Signed-off-by: Tim Schmielau <tim@physik3.uni-rostock.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>