kbuild: Use POSIX headers for ntoh functions
From: Benno <benjl@cse.unsw.edu.au>
When compiling Linux on Mac OSX I had trouble with scripts/sumversion.c.
It includes <netinet/in.h> to obtain to definitions of htonl and ntohl.

On Mac OSX these are found in <arpa/inet.h>. After checking the POSIX
specification it appears that this is the correct place to get
the definitons for these functions.

(http://www.opengroup.org/onlinepubs/009695399/functions/htonl.html)

Using this header also appears to work on Linux (at least with
Glibc-2.3.2).

It seems clearer to me to go with the POSIX standard than implementing
#if __APPLE__ style macros, but if such an approach is preferred I can
supply patches for that instead.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>