[PATCH] ipr: Use cancel all instead of abort task
Currently, ABORT_TASK does not work properly on ipr adapters. It results in
timeouts and eh ends up getting escalated to eh_host_reset to recover.
The adapters are supposed to be fixed to properly handle this at some
point in the future, but until that updated microcode is available,
we need to send a cancel all to the device instead.

Signed-off-by: Brian King <brking@us.ibm.com>
Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>