[IA64-SGI] bte.c: kill expression as lvalue warning
A recent patch caused a warning about not using expressions as lvalues
to crop up in bte.c (or maybe it's just that I'm using gcc-3.4.1 now).
This patch fixes it by creating a temporary to store the register whose
address we want to get and stuff into the per-bte info structure.

Signed-off-by: Jesse Barnes <jbarnes@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>