kbuild: Remove check for undefined symbols in vmlinux
Russell King reported that some gcc / bin-utils combination may result in
undefined symbols in vmlinux and implemented a check for that.
Though the cause is very valid this check has proved to be a real pain
for other users, especially sparc and um.
Short term a similar check is asked to be implemented in arch/arm*/Makefile
and long-term to bail out if too old gcc / bin-utils is used.
Long term plan awaits a new gcc / bin-utils release.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>