[ACPI] ACPICA 20040816 update from Bob Moore
Designed and implemented support within the AML interpreter
for the so-called implicit return.  This support returns
the result of the last ASL operation within a control
method, in the absence of an explicit Return() operator.
A few machines depend on this behavior, even though it
is not explicitly supported by the ASL language.  It is
optional support that can be enabled at runtime via the
acpi_gbl_enable_interpreter_slack flag.

Removed support for the PCI_Config address space from the
internal low level hardware interfaces (acpi_hw_low_level_read
and acpi_hw_low_level_write).  This support was not used
internally, and would not work correctly anyway because
the PCI bus number and segment number were not supported.
There are separate interfaces for PCI configuration space
access because of the unique interface.
acpica-unix-20040816.patch

AE_CODE_AML_MAX fix from Bjorn Helgaas