[PATCH] ppc64: Don't call scheduler on offline cpu
When taking a cpu offline, once the cpu has been removed from
cpu_online_map, it is not supposed to service any more interrupts.  This
presents a problem on ppc64 because we cannot truly disable the
decrementer.  There used to be cpu_is_offline() checks in several scheduler
functions (e.g.  rebalance_tick()) which papered over this issue, but these
checks were removed recently.  So with recent 2.6 kernels, an attempt to
offline a cpu can result in a crash in find_busiest_group().  This patch
prevents such crashes.

Signed-off-by: Nathan Lynch <nathanl@austin.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>