[PATCH] Fix x86-64 singlestep through sigreturn system call
This patch just directly clones Davide Libenzi's i386 code for x86-64 in
both 64-bit and 32-bit cases.  With this, the behavior of single-stepping
all system calls is consistent.  

The syscall exit tracing caused by TIF_SINGLESTEP is superfluous in the
case of sysret returns, but harmlessly so (since continuing afterward with
PTRACE_CONT will have cleared TF as well as TIF_SINGLESTEP).  I figured
that little bit of extra processing in the single-step case was better than
adding code to ignore the flag in the sysret case.

Signed-off-by: Roland McGrath <roland@redhat.com>
Cc: <ak@muc.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>