[PATCH] Add BLIST_INQUIRY_36 to all USB blacklist entries
This patch adds the BLIST_INQUIRY_36 flag to all the SCSI blacklist
entries for USB devices.  While it may not be strictly necessary for all
of them, it doesn't hurt: Since the usb-storage driver doesn't use any of
the INQUIRY data after the first 36 bytes, there's no reason to try
reading any more of it.  And some devices crash when we try to read more,
even though they advertise that more bytes are available.  The usb-storage
driver does try to set the flag automatically, but the blacklist entries
override that setting.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>