[IPV6]: Add option to copy DSCP in decap in ip6_tunnel.
Here is a patch that allows the copying of the DSCP during decapsulation
for ip6_tunnel.  I've made it a separate option from the one that
determines the copying during encapsulation since the DSCP processing
may be asymmetric.  It also means that we preserve compatibility should
anyone be relying on the current behaviour.

inet_ecn.h might appear to be an odd place for ipv6_copy_dscp, but
I couldn't put it in dsfield.h since I want to use ipv6_get_dsfield
in inet_ecn.h later on.  The other alternative would be to define
INET_ECN_MASK in dsfield.h.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>