[PATCH] ES7000 subarch update
The patch below implements an algorithm to determine an unique GSI override
for mapping GSIs to IO-APIC pins correctly.  GSI overrides are required in
order for ES7000 machines to function properly since IRQ to pin mappings
are NOT all one-to-one.  This patch applies only to the Unisys specific
ES7000 machines and has been tested thoroughly on several models of the
ES7000 line.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>