[PATCH] ppc32: 85xx spurious interrupt bug
The Spurious interrupt assignment for OpenPIC was the constant,
OPENPIC_VEC_SPURIOUS plus the offset passed in.  Due to bugs in some
OpenPICs, the end value needed to be 255 on those systems.  However, the
constant was then chosen to be 239, because offset on most systems is 16.
The 85xx has an offset of 64, however, and so this assignment caused
lock-ups during boot.  The proposed solution is to declare
OPENPIC_VEC_SPURIOUS to be 255, and use that value directly, instead of
adding the offset to it.

Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>