[PATCH] dio fine alignment and pages in io
Here is the fix for dio_new_bio() BUG() with fsx tests.  The problem is, I
changed dio_pages_in_io calculation to be exactly the number of pages we
needed to map the user-buffer to do the IO.  But with fine-grain alignment
(non fs-blocksize), we may need 2 more pages (ZERO_PAGE) to zero out first
and last block if needed.

Here is the bug info.

http://bugme.osdl.org/show_bug.cgi?id=3447

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>