[IA64-SGI] sn2: serialize access to PROM chips
If we read and write the PROM chips at the same time, as might happen at boot
when salinfo extracts MCA records and a user is checking the PROM revision
in /proc/sgi_prominfo, an MCA might occur, since the PROM chips can't be
accessed that way.  This patch fixes the problem for systems with new PROMs
(>= 3.50) by using the SAL to do PROM reads.

Signed-off-by: Jesse Barnes <jbarnes@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>