[PATCH] PPC64: RPA dynamic addition/removal of PCI Host Bridges
From: John Rose <johnrose@austin.ibm.com>

The following patch implements the ppc64-specific bits for dynamic (DLPAR)
addition of PCI Host Bridges.  The entry point for this operation is
init_phb_dynamic(), which will be called by the RPA DLPAR driver.

Among the implementation details, the global number aka PCI domain for the
newly added PHB is assigned using the same simple counter that assigns it at
boot.  This has two consequences.  First, the PCI domain associated with a PHB
will not persist across DLPAR remove and subsequent add.  Second, stress tests
that repeatedly add/remove PHBs might generate some large values for PCI
domain.  If we decide at a later point to hash an OF property to PCI domain
value, this can be easily fixed up.

Also, the linux,pci-domain property is not generated for the newly added PHBs
at the moment.  Because there doesn't seem to be an easy way to dynamically add
single properties to the OFDT, and because the userspace dependency on this
property is being questioned, I've ignored it for now.  If we decide on a
solution for this at a later point, it can also be easily fixed up.

Signed-off-by: John Rose <johnrose@austin.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>