[PATCH] USB: USB CDC OBEX driver
Also, as the full patch isn't going in, can you please apply this tiny
part of it? Somehow the header descriptor was omitted from the CDC ACM
driver, but it's present on my phone (thus giving me the "ignoring extra
header" error when it's plugged in) and in the CDC spec (section 5.2.3.1).


Signed-off-by: Alex Kanavin <ak@sensi.org>
Signed-Off-By: Oliver Neukum <oliver@neukum.name>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>