[ARM] Sanitise Footbridge machine class.
Footbridge was suffering from a little lack of care and attention;
it still had the nasty arch.c file with all the associated #ifdef
gross-ness that entailed.

Re-jig footbridge support so that each machine type contains all
the necessary support code, with a separate common implementation
which they all share.