[ACPI] ACPICA 20040922 from Bob Moore
Signed-off-by: Len Brown <len.brown@intel.com>

Fixed a problem with the implementation of the LNot()
operator where "Ones" was not returned for the TRUE
case. Changed the code to return Ones instead of (!Arg)
which was usually 1. This change affects iASL constant
folding for this operator also.

Fixed a problem in acpi_ut_initialize_buffer where an
existing buffer was not initialized properly -- Now zero
the entire buffer in this case where the buffer already
exists.

Changed the interface to acpi_os_sleep from (UINT32
Seconds, UINT32 Milliseconds) to simply (ACPI_INTEGER
Milliseconds). This simplifies all related code
considerably. This requires changes/updates to all OS
interface layers (OSLs.)

Implemented a new external interface,
acpi_install_exception_handler, to allow a system exception
handler to be installed. This handler is invoked upon
any run-time exception that occurs during control method
execution.

Added support for the DSDT in acpi_tb_find_table. This
allows the DataTableRegion() operator to access the local
copy of the DSDT.