[PCMCIA] Fix PCMCIA behaviour on resume with different card.
PCMCIA checks the card CIS against its cached copy.  If it finds
that the card does not match, it destroys the bindings with existing
drivers, issues an remove event followed by an insert event.  However,
ds.c delays the remove event by 100ms, so cardmgr sees the insert
before remove.  It thereby ignores the new card.

Also, we ended up leaving the fake CIS intact, so the new card appears
to be described by the fake CIS.  Destroy the fake CIS in addition to
the CIS cache.