[PATCH] lsm: fix send_sigurg mediation
Stephen Smalley notes that send_sigurg isn't mediated by LSM in the same
manner as send_sigio.  Patch below is a slight modification of Stephen's
original patch.  It moves the security_file_send_sigiotask() hook into the
sigio_perm().  The hook's fd and reason arguments are replaced with the
signum.  sigio_perm() and it's callers are updated to pass the signum
through to the hook.  In send_sigio case, the signum is simply fown->signum
or SIGIO when signum is 0, however in send_sigurg the kernel doesn't use
fown->signum, it always sends SIGURG.

From: Stephen Smalley <sds@epoch.ncsc.mil>
Signed-off-by: Chris Wright <chrisw@osdl.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>