[PATCH] md: make retry_list non-global in raid1 and multipath
Both raid1 and multipath have a "retry_list" which is global, so all raid1
arrays (for example) us the same list.  This is rather ugly, and it is simple
enough to make it per-array, so this patch does that.

It also changes to multipath code to use list.h lists instead of
roll-your-own.

Signed-off-by: Neil Brown <neilb@cse.unsw.edu.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>