[PATCH] I2C: New LM63 chip driver
This is a new I2C chip driver named lm63, which supports National
Semiconductor's LM63 hardware monitoring chip. The LM63 is similar to
the LM86 (which we do support through our lm90 driver) with additional
monitoring and control of a single fan. It is obviously aimed at the CPU
and high-end graphics adapters markets. This new driver is very similar
in nature to the lm83 and lm90 driver so I don't have much to add.

I developed the driver on a request from Hard Data Ltd. which are using
Tyan S4882 boards where four LM63 chips are used (one for each CPU).
Tyan provided remote access to a test system so that I could test my
driver on real LM63 chips.


Signed-off-by: Jean Delvare <khali@linux-fr.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>