Kbuild: avoid backup localversion files
Avoid including as localversion-files the *~ files, i.e. backup files. If I
have localversion-a and localversion-a~, I don't want both to be used. Nor I
want to use localversion*~ anyway.

Don't code that as $(wildcard localversion*[^~]) as that would exclude
"localversion" from the wildcard expansion result, because it requires at
least one character after localversion to exist in the name file. I.e.,

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade_spam@yahoo.it>