[Bluetooth] Support responses with zero SCID
According to the Bluetooth specification there is no need to fill in
the SCID on a failing response. Without a SCID the L2CAP core can't
find the corresponding request for it. In the case of the zero SCID
the signal identifier should be used to match the response to its
request.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>