[PATCH] make cdu31a work on at least one system
I've got a Sony CDU33A drive with COR334 controller.  The Linux cdu31a
driver was not updated for 2.6 kernel so it does not work.

Here are patches that try to make the driver working with 2.6
kernel:
  - fix the timeout values in header file
  - Make the driver work in 2.6.X
  - Added workaround to fix hard lockups on eject
  - Fixed door locking problem after mounting empty drive
  - Set double-speed drives to double speed by default
  - Removed all readahead things - not needed anymore

It does work on my system. I also know that it's still broken - it uses 
cli(), MODULE_PARM and it's also not very fast (I _never_ reached full
300KB/s with it, but I know that it's possible in Windows) and probably 
many other things (I'm new to Linux kernel) - so I'm waiting for comments.