[PATCH] fbcon: Add box drawing glyphs to 6x11 font
From: David Lazar <david@panzer.utcluj.ro>

I've tried to use the 6x11 (mac console) font on the framebuffer console (on
the x86 architecture), and I've noticed that the font didn't have the correct
box drawing glyphs.  As a result I've used as a model the 8x16 font and edited
the glyphs for all 256 US ASCII characters, so that they display correctly.

Signed-off-by: Antonino Daplas <adaplas@pol.net>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>