[PATCH] ppc64: Fix for cpu hotplug + NUMA
It was possible to hotplug add a new cpu that was larger than numnode.
This meant we had a cpu with a node id that didnt have any NODE_DATA()
backing it.

The following patch catches this in numa_setup_cpu and forces the node to
0.  We also make a pass through all cpus at boot to look for the maximum
node id - this catches the case at boot.  As for hotplug, since the node
doesnt have any memory backing it (otherwise the node would have been
onlined) then we can assume the machine isnt configured for performance and
just force the cpu into node 0.

Signed-off-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>