[PATCH] uml: Don't delay segfaults
From: Bodo Stroesser <bstroesser@fujitsu-siemens.com>

This one covers the fact, that the SIGSEGV signal, which is created by
force_sigsegv() in case of an error in handle_signal(), is not delivered to
the user immediately.  In the worst case it even could be masked if a
sigprocmask() systemcall follows immediately after return from kernel.  The
patch is relevant for other architectures, too.

Signed-off-by: Jeff Dike <jdike@addtoit.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>