update the fc_transport_class to use a workqueue instead of a timeout
The amount of work that has to be done in the timeout routines is really
a bit much if there's a large number of LUNS.  Plus the
device_for_each_child needs user context to get the bus semaphore, so
the solution is to migrate them from a timer to delayed work.

There's still a race here in that the timer may still be ticking when
the device is destroyed ... To fix this, I think we may need to
introduce a destroy callback to the transport class.

Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>