MTD: Provide XIP support for Intel flash chips.
This allows for MTD support to be used on flash memory which is also used
for XIP purposes, either XIP kernel or XIP userspace.  The whole idea is
to relocate functions actually modifying the flash state away from array mode
to ram and run them with interrupt disabled.  When the flash needs some time
to complete a certain operation, we poll the processor for pending (but still
masked) interrupts, and when they occur we suspend the flash operation in order
to unmask interrupts and let the system run again.

Signed-off-by: Nicolas Pitre <nico@cam.org>
Signed-off-by: David Woodhouse <dwmw2@infradead.org>