[NET]: Add alloc_skb_from_cache.
This serves two purposes: firstly, we like to allocate page-sized skbs
as this means we zero-copy transfer of network buffers between guest
operating systems. Secondly, it enables us to have a cache of pages
that have been used for network buffers that we can be more lax about
scrubbing when they change VM ownership (since they could be sniffed on
the wire).

Signed-off-by: David S. Miller <davem@davemloft.net>