[PATCH] vmscan: ignore swap token when in trouble
The token-based thrashing control patches introduced a problem: when a task
which doesn't hold the token tries to run direct-reclaim, that task is told
that pages which belong to the token-holding mm are referenced, even though
they are not.  This means that it is possible for a huge number of a
non-token-holding mm's pages to be scanned to no effect.  Eventually, we give
up and go and oom-kill something.

So the patch arranges for the thrashing control logic to be defeated if the
caller has reached the highest level of scanning priority.


Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>