[PATCH] smbfs: Bug #3758 - Broken symlinks on smbfs
The patches attached here fixes the BUG #3758 - "Broken symlinks on smbfs
with 2.6.10-rc[12]".  There are two patches, one to be applied over version
2.6.9 and the other over 2.6.10-rc2.

The old utilities (and the old driver) uses uid=0, gid=0, dmask=0 and
fmask=0 to flag the lack of this parameters on the mount command line.
When the user do not specify the uid, gid, fmask or dmask, the current
driver will assign gid=root, gid=root, dmask=755, fmask=755.  This behavior
is similar to the old 2.x samba versions.

To make the driver see the permissions and ownership assigned on the
server, "smbmount" and "smbmnt" utilities must be patched.  The patches is
already available on the attached patches at Bug #3330 or Samba Bug #999.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>