[ARM PATCH] 2279/1: [PATCH] SIMpad: Add a mq200 device to the platform bus
Patch from Holger Hans Peter Freyther

The SIMpad uses the MediaQ 200 framebuffer device. There is no driver in the vanilla kernel for that device. But adding the device to the platform bus makes it possible to just drop the mq200 driver into the kernel and the display will work.
Please consider applying the patch.



Signed-off-by: Holger Hans Peter Freyther
Signed-off-by: Russell King