[PATCH] sys_set/getpriority PRIO_USER semantics fix and optimisation
This change brings the semantics equivalent to 2.4 and also to what the man
page says; Also optimises by avoiding unneeded lookup in uid cache, when
who is same as the current->uid.

sys_set/getpriority is rewritten in 2.5/2.6, perhaps while transitioning to
the pid maps.  It has now semantical bug, when uid is zero.  Note that akpm
also fixed refcount leak and locking in the new functions in changeset
http://linus.bkbits.net:8080/linux-2.5/cset@1.1608.10.84

Signed-off-by: <pmeda@akamai.com>
Acked-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>