[IA64-SGI] Add support for a future SGI chipset (shub2) 2of4
This patch adds the addresses of shub2 MMRS to the shub_mmr
header file. During boot, a SAL call is made to determine the
type of the shub. Platform initialization sets the appropriate
MMR addresses for the platform.

A new macro (is_shub1() & is_shub2()) can be used at runtime to
determine the type of the shub.

Signed-off-by: Jack Steiner <steiner@sgi.com>