[PATCH] highmem.c: fix bio error propagation
Found a subtle bug that caused mount errors on a SATA drive with
barriers on reiser and ext3, where it should have recovered and just
turned off barriers. The problem is that the EOPNOTSUPP error isn't
being propagated properly to the bounced bio. This patch fixes that by
correctly passing error all the way down.

Signed-off-by: Jens Axboe <axboe@suse.de>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>