[IA64] Clear all corrected records as they occur
Because MCA events are not irq safe, they cannot be logged via
salinfo_decode at the time that they occur.  Instead kernel salinfo.c
runs a timer every few minutes to check for and to clear corrected MCA
records.  If a second recoverable MCA occurs on the same cpu before
salinfo_decode has cleared the first record then OS_MCA reads the
record for the first MCA from SAL, which passes invalid data to the MCA
recovery routines.

This patch treats all corrected records the same way, by clearing the
records from SAL as soon as they occur.  CMC and CPE records are
cleared as they are read.  Recoverable MCA records are cleared at the
time that we decide they can be corrected.  If salinfo_decode is not
running or is backlogged then we lose some logging, but that has always
been the case for corrected errors.

Signed-off-by: Keith Owens <kaos@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>