[PATCH] USB: EHCI "park" mode disabled
This changes the default initialization of the EHCI "park" mode so
that silicon which supports it (NF2, NF3, ALI, GeneSys, ...) will not
use it unless explicitly told to do so by a (new) module parameter.

This is a workaround for some problems observed on some NF2 systems:

    - Throughput ("hdparm -tT") is lower than expected with recent high
      performance drives (Maxtor) ... disabling "park" increases it by
      about 2 MByte/sec, but it's still much slower than expected.
      (USB analyser traces should be informative here.)

    - Some data corruption observed on reads from drives using an
      ALI storage adapter ... disabling "park" stops the corruption.
      (Strongly suggestive of hardware or peripheral firmware bugs;
      multiple back-to-back bulk-IN packets should work just fine.)

The "don't use park mode" workaround will at most reduce USB (and PCI)
throughput slightly on systems that work as expected, but some of those
can re-enable the "park" mode.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>