SCSI: add queue_type entry in sysfs
This adds an extra attribute to tell you what type of queueing the
driver is using: none, simple or ordered.  If the driver supplies the
change_queue_type API, you can also alter this (which would allow the
turning on or off of TCQ).

I also fixed the change_queue_depth not to allow the user to go below
one.

Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>