[PATCH] make microcode text less confusing
Red Hat got some confused customers due to this message. The confused
user case is when they update the BIOS and all of a sudden we have "no
suitable data" yet we did before. We (Arjan van de Ven) thus changed it
to "No new microcode" which is much much clearer.

Signed-off-by: Alan Cox <alan@redhat.com>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>