[IA64] contig.c: fix bug in DMA-zone initialization
Below is a patch that fixes count_dma_pages().  Without this fix, only
memory areas entirely below MAX_DMA_ADDRESS are counted as DMA-pages.
While it doesn't really make a difference in practice (MAX_DMA_ADDRESS
is equal to ~0UL on zx1), it's clearly not what was intended to be
there.  I found the bug only because I briefly experimented with
lowering MAX_DMA_ADDRESS.

Tony> The discontig.c version of this code already gets this right.
 
Signed-off-by: David Mosberger-Tang <davidm@hpl.hp.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>