[PATCH] make gconfig work with gtk-2.4
I need this to make gconfig work under gtk-2.4.  Without this, it just
coredumps.  There is some problem with pixmap creation/usage from XPM in
the way it is done in gconf, so I just added some stock icons.  It is even
prettier..;)

Could someone test this still works on gtk-2.0 or 2.2 ?

Changes:

- change the wiget class 'button' in glade files to something known to
  glade (GtkToolButton)
- use 'stock-id' property for toolbar buttons instead of "stock_pixmap"
- change unknown signal "pressed" to "clicked"
- remove manual setting of icons in gconf.c

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>