[PATCH] Ext[23]: apply umask to symlinks with ACLs configured out
Keith Young <stripyd@stripydog.com> has reported that when ACLs are not
compiled in, the default implementation of ext[23]_init_acl applies the
umask to all new files, including symlinks, which is wrong.  In this case
the VFS already takes care of applying the umask when needed, so ext2 and
ext3 need not bother about it.  Remove the superfluous statements.

Signed-off-by: Andreas Gruenbacher <agruen@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>