[IA64] Stagger the addresses of the pernode data structures to minimize cache aliasing.
Allocation of pernode structures in find_pernode_space() does not
properly stagger the alignment of the pgdats. This causes
aliasing of the structures in the L3 caches, ie. the same fields
in pgdat structures for multiple nodes will index to same cache
index in the L3. 

If a process is allocating a huge amount of space & many nodes must 
be scanned before finding a node with available space, allocation 
of a pages is significantly slowed by excessive cache misses.

By properly staggering the locations of the pgdat structures, allocation
times on insanely large systems is dramatically improved. On a 256 node
512GB system, allocation of 450 GB by a single process was reduced 
from 1510 sec to 220 sec - a 7X improvement.

Aside from wasting a trivial amount of space, I don't see any 
downside to staggering the allocation by 1 cacheline per node.

	wasted space
		bytes = N * (N-1) * 64

	For 64 node system
		wasted bytes = ~256K

Signed-off-by: Jack Steiner <steiner@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>