[ARM PATCH] 2358/1: Add IXP46x I2C platform device
Patch from Deepak Saxena

The IXP46x chips from Intel include the same I2C controller as found
on the IOP3xx.  This patch adds a hook to the platform-level init code
to add that device to the device tree if running on a 46x CPU. A
separate patch has been sent upstream via Greg KH for the changes to the
I2C driver to support both IOP and IXP systems.

While at it, we fix a minor issue with the code where we were sending 
the wrong ptr to platform_add_devices(). I am not sure how it ever 
worked...

Signed-off-by: Deepak Saxena
Signed-off-by: Russell King