[PATCH] sched: active_load_balance() fixlet
There is a small problem with the active_load_balance() patch that Darren
sent out last week.  As soon as we discover a potential 'target_cpu' from
'cpu_group' to try to push tasks to, we cease considering other CPUs in
that group as potential 'target_cpu's.  We break out of the
for_each_cpu_mask() loop and try to push tasks to that CPU.  The problem is
that there may well be other idle cpus in that group that we should also
try to push tasks to.  Here is a patch to fix that small problem.  The
solution is to simply move the code that tries to push the tasks into the
for_each_cpu_mask() loop and do away with the whole 'target_cpu' thing
entirely.  Compiled & booted on a 16-way x440.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>