[PATCH] idle thread preemption fix
The early bootup stage is pretty fragile because the idle thread is not yet
functioning as such and so we need preemption disabled.  Whether the bootup
fails or not seems to depend on timing details so e.g.  the presence of
SCHED_SMT makes it go away.

Disabling preemption explicitly has another advantage: the atomicity check
in schedule() will catch early-bootup schedule() calls from now on.

The patch also fixes another preempt-bkl buglet: interrupt-driven
forced-preemption didnt go through preempt_schedule() so it resulted in
auto-dropping of the BKL.  Now we go through preempt_schedule() which
properly deals with the BKL.

Signed-off-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>