[PATCH] fbdev: rivafb should recognize NF2/IGP
I got tired of not seeing the boot time penguin on my Shuttle SN41G2, and
not having a decently large text display when I bypass X11.  XFree86 says
it's "Chipset GeForce4 MX Integrated GPU", and the kernel driver has hooks
for this chip ID although it doesn't have a #define to match.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>