[PATCH] 3c59x: enable power management unconditionally
The problem is that under the new PCI driver model, cards are required to
always restore state and call pci_enable_device() on resume.  So the patch
changes the driver to do its restore state calls unconditionally (it used
to only do them when it was configured for wake-on-lan) and adds a call to
pci_enable_device().

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>