[PATCH] pcmcia: device model integration can only be submitted under GPL, part 2
As discussed previously, my integration of ds.c and cs.c with the driver model
can and will only be available under the GPL, as it's too much derived of
other buses' implementation of integration with the driver model.

cs_internal.h did only contain the MPL header before - I contacted Dave Hinds
because of this, and as far as he can tell, it was just an oversight that this
was not marked as dual-licensed as the other files are.

Signed-off-by: Dominik Brodowski <linux@brodo.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>