[PATCH] pcmcia: remove client_attributes
Remove the "Attributes" entry of struct client_t -- _all_ in-kernel drivers
did set it to INFO_IO_CLIENT | INFO_CARD_SHARE anyway, and the pcmcia core
didn't care at all what this value was set to.

Also, remove the setting of the respective request field in all in-kernel
drivers -- the request structure is kept the same, though, to keep external
drivers compiling.

Signed-off-by: Dominik Brodowski <linux@brodo.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>