[PATCH] InfiniBand/ipoib: move structs from stack to device private struct
Move the gather list and work request used for posting sends from the stack in
ipoib_send() to the private structure.  This reduces the stack usage for the
data path function ipoib_send() and may speed things up slightly because we
don't need to initialize constant members of the structures.

Signed-off-by: Roland Dreier <roland@topspin.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>