[PATCH] fix coredump_wait deadlock with ptracer & tracee on shared mm
In the oddball situation where one thread is using ptrace on another thread
sharing the same mm, and then someone sharing that mm causes a coredump,
there is a deadlock possible if the traced thread is in TASK_TRACED state. 
It leaves all the threads sharing that mm wedged and permanently
unkillable.  This patch checks for that situation and brings a thread out
of TASK_TRACED if its tracer is part of the same coredump (i.e.  shares the
same mm).  It's not pretty, but it does the job.

Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>