[PATCH] fix xenU kernel crash in dmi_iterate
In unprivileged Xen domains, all that __ioremap() does is a "return NULL",
which causes dmi_iterate() to crash the kernel at boot time.

This trivial check bails dmi_iterate() out of the loop when it finds that
the ioremap() returned a NULL pointer.

(akpm: this fix goes beyond xen - ioremap can return NULL)

Signed-off-by: Rik van Riel <riel@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>