[Bluetooth] Fix too many keys pressed error
This patch fixes the problem of keys repeating when too many keys
are pressed at the same time (e.g. when typing quickly). This
often results in "mount" becoming "mouount".

It seems that Bluetooth keyboards send a HID report with the keys
all set to 0x01 if too many keys were pressed at the same time.
This confuses the previous report handling logic and now these
reports are ignored.

Signed-off-by: Juha Yrjölä <juha.yrjola@iki.fi>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>