[PATCH] e100 locking up netconsole.
From: Steven Rostedt <rostedt@goodmis.org>

If the transmit buffer of the e100 overflowed, then the system would hang.
This was caused because the e100 driver would stop the queue, and
netpoll_send_pkt in netpoll.c would then loop forever.  This is because the
e100 net_poll would never start the queue again after the transmits have
completed.

For those that use the e100 and netconsole, all you need to do is a sysreq
't' to lock up the system.

Acked-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>