[ARM PATCH] 2442/1: Simplifying NODES_SHIFT
Patch from Marc Singer

The special case for the Sharp LH processors is unnecessary.  A macro
override makes it cleaner and concentrates the change where it ought
to be.  The default in include/asm-arm/numnodes.h means that only
platforms that care to change the default need to do anything.

Signed-off-by: Marc Singer
Signed-off-by: Russell King