[PATCH] kbuild: no redundant srctree in tags file
Avoid cluttering the tags/TAGS generated file with $(srctree) in the paths
if this is not needed.

This has two advantages:

- Saving about 20M on the size of the resulting tags file (which are used
  currently to store the absolute path of the file names rather than the
  relative one) when KBUILD_OUTPUT is not set.

- Keeping the tags file valid when the directory is renamed.

No change is done for who does make tags O=..., if this is wanted (I would
find that incommodous and non-typical for a developer, but anyway I've not
ruined functionality in that case).

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade@yahoo.it>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Sam Ravnborg <sam@ravnborg.org>