[PATCH] SELinux: fix selinux_inode_setattr hook
This fixes the selinux_inode_setattr hook function to honor the ATTR_FORCE
flag, skipping any permission checking in that case.  Otherwise, it is
possible though unlikely for a denial from the hook to prevent proper
updating, e.g.  for remove_suid upon writing to a file.  This would only
occur if the process had write permission to a suid file but lacked setattr
permission to it.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>
Signed-off-by: James Morris <jmorris@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>