[PATCH] update aty128fb sleep/wakeup code for new powermac changes
This patch updates aty128fb power management code to the changes of the
powermac sleep mecanism.  It makes the driver use the new hook for early
wakeup, adds the call to the arch code indicating wether it can wakeup the
chip, etc...  This patch shouldn't break non-ppc, but this should be
tested.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>