[IPV4]: Make loopback idev stick around.
As it is when loopback_dev loses all of its IPv4 addresses its
corresponding idev will be destroyed.  Unfortunately as of last
August route.c relies on the loopback idev to kill references
to other idev objects.

The end result is that when you do ip a f dev lo, unregistering
other devices will hang until those dst objects referring to
their idev objects die of natural causes.  Of course this may
never happen if the processes holding those references get
dead-locked by invoking an operation that takes the RTNL.

A simple solution is to make sure that loopback's idev sticks
around all the time.

Incidentally this also fixes the setting of some flags on the
loopback idev object as currently the code that does it won't
be called if you add the addresses to lo after bring it up.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>