[PATCH] sata_qstor: new basic driver for Pacific Digital
This is a new basic libata SATA driver
for the Pacific Digital QStor SATA/RAID card.

It features ordinary per-drive SATA with single-request DMA
for R/W commands, and PIO for everything else.

It currently does not implement any of the hardware RAID support,
nor hot-plug, nor the tagged/native command-queuing features.

On the other hand, it is small and simple as a result.

Signed-off-by: Mark Lord <mlord@pobox.com>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>