[ARM PATCH] 2460/1: fix up resource usage on locomo
Patch from John Lenz

Add the list of devices on the locomo chip, and change around
how resources and struct resource are used.  There is only one
struct resource for the entire locomo, but each driver will
call request_mem_region on the pieces it is using.
Secondly, add a few helper functions to locomo.c to control
GPIOs and DAC.

Signed-off-by: John Lenz
Signed-off-by: Russell King