[PATCH] ppc32: Don't create .tmp_gas_check
This changes how the ppc32 'checkbin' target works.  Previously we did all
of the tests using variables which would be evaluated for any and every
make target.  This meant that 'make tags' for example would leave behind a
'.tmp_gas_check' file even though we didn't actually compile up anything,
and would get in the way of diffs.  By moving all of this logic directly
into the 'checkbin' target, we only test gcc/gas versions when we're going
to compile.

Signed-off-by: Olaf Hering <olh@suse.de>
Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>