[PATCH] usb/io_edgeport: remove interruptible_sleep_on_timeout() usage
On Tue, Feb 01, 2005 at 11:43:08AM -0800, Nishanth Aravamudan wrote:
> This should fix the behavior of the previous patch (probably not noticed yet).
> The wake_up*() was not matched properly in the first patch (fixed below). Please
> consider reverting the previous patch and applying this one instead.

Directly use wait-queues to remove remaining callers of
interruptible_sleep_on_timeout(). Signals are not checked (except in one
case) in the current code, so interruptible should not be necessary.
Modify the wake_up*() calls to match the wait-queue usage. There were
some naming conflicts, which I tried to resolve appropriately. The final
replacement is within a #if 0 / #endif section of code, but in case that
code is ever used again, I would prefer it had the correct interface :)

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>