[PATCH] OSS Support for AC97 low power codecs
This is a resend of a patch that has been applied to 2.4.  The low power
codec functionality has also now been included in ALSA.

It checks the codec ID before doing an AC97 register reset.  This allows
the kernel to support low power codecs that are powered down by a reset
command.  This patch also fixes some other minor issues.

Changes:-

- Added AC97_DEFAULT_POWER_OFF to ac97_codec_ids[]

- ac97_probe now checks hardwired codec ID's before sending a reset

- Added support for WM9713

- Moved the codec specific inits after the mixer setup as some init

- tings were being clobbered.

- Added extra check so that default_digital_ops doesn't overwrite a valid
  codec_ops.  (SPDIF)

Signed-off-by: Liam Girdwood <liam.girdwood@wolfsonmicro.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>