[PATCH] Add struct request end_io callback
This is needed for several things, one in-tree user which I will introduce
after this patch.

This adds a ->end_io callback to struct request, so it can be used with
async io of any sort.  Right now users have to wait for completion in a
blocking manner.  In the next iteration, ->waiting can be folded into
->end_io_data since it is just a special case of that use.

From: Peter Osterlund <petero2@telia.com>

The problem is that the add-struct-request-end_io-callback patch forgot to
update pktcdvd.c.  This patch fixes it.

Signed-off-by: Jens Axboe <axboe@suse.de>
Signed-off-by: Peter Osterlund <petero2@telia.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>