[PATCH] Fix 1-Wire Dallas in bigendian machines
I've been testing the 1-Wire Dallas in a bigendian machine (through a GPIO)
and I've found some problems that can easily addressed with the provided
patch.  (inline at the end of the message).

I have a question about the implementation of w1_smem.
In the line 90 of drivers/w1/w1_smem.c.
  for (i = 0; i < 9; ++i)
     count += sprintf(buf + count, "%02x ", ((u8 *)&sl->reg_num)[i]);
I don't see why this loop is execute 9 times when the provided reg_num
is 8 bytes long. I don't understand the purpose of the last byte.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>