[PATCH] posix-timers: fix posix-timers signals lock order
The posix-timers code establishes the locking order that k_itimer locks are
outside siglocks.  However, when the signal code calls back into the
posix-timers code to reload a timer after its signal is dequeued, it holds a
siglock while calling do_schedule_next_timer, which gets a timer lock.

I'm not sure there is any deadlock scenario possible using the real-time POSIX
timers, because of the intricate arrangement of timer firing and resetting
synchronization.  But with the new CPU timers code, this deadlock pops up
right away.  Dropping the siglock here certainly doesn't hurt in the real-time
timer cases, and it really seems like the right thing here to keep the locking
details in the interface between signals and posix-timers code comprehensible.

Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>