[PATCH] ppc64: "invert" dma mapping routines
This patch "inverts" the PPC64 dma mapping routines so that the pci_ and
vio_ ... routines are implemented in terms of the dma_ ... routines (the
vio_ routines disappear anyway as noone uses them directly any more).

The most noticable change after this patch is applied will be that the
flags passed to dma_alloc_coherent will now be honoured (whereas they were
previously silently ignored since we used to just call
pci_alloc_consistent).

Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>