[PATCH] x86: memset the i386 numa pgdats in arch code
The next patch in this series will remove the arch-independent clearing of the
pgdat's.  This first patch removes the i386 dependency on that behavior.

The new i386 function, remapped_pgdat_init() takes care of initializing the
pgdats which are finally mapped after paging_init() is done.  The
zone_sizes_init() call has to occur after the pgdat clearing.

zone_sizes_init() is currently called from the end of paging_init(), because
that's the first place where the pgdats could have been zeroed.  However,
zone_sizes_init() really doesn't have anything to do with paging, and probably
shouldn't be in paging_init().

Moving this call into setup_memory() allows the declaration of
zone_sizes_init() to change files as well, which means a net removal of one
#ifdef.  It also provides a handy place to put the new function, far away from
the paging code that it really has nothing to do with.  Moving files required
only using highend_pfn inside of the HIGHMEM ifdef, but this saves a line of
code anyway.

Fixes from: Yasunori Goto <ygoto@us.fujitsu.com>
Signed-off-by: Dave Hansen <haveblue@us.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>