[PATCH] nfsd4: fix delegation filp sharing
The st_vfs_file pointer is set when the open that originally got the delegation
is closed, and then in release_deleg we attempt to distinguish between the case
when the original open has been closed and when it hasn't.  The resulting logic
is complicated and buggy.

It's much simpler just to keep a reference count on the struct file.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Neil Brown <neilb@cse.unsw.edu.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>