[PATCH] nvidiafb: Add update framebuffer driver for nVidia chipsets
Because nVidia keeps pumping out new graphics chipsets, it is becoming
harder to add support for these with the old rivafb code.  Currently, rivafb
can properly support NV_ARCH_20 chipsets and older.

Instead of rewriting rivafb to support the latest chipsets, I've decided to
write a new driver, called nvidiafb.  The aim is to closely follow Xorg
development.

Currently, this driver is based on the most recent CVS Xorg nv driver. Main

  - console acceleration for all chipsets
  - uses DMA instead of PIO
  - better LCD/digital output support
  - better monitor detection
  - support for Riva128 will be dropped as it cannot do DMA, rivafb
    will remain as the driver for this chipset

It should work with Xorg/XFree86 nv driver, but as with rivafb, is not
compatible with the proprietary nvidia driver.

Once the code becomes stable, rivafb code will be trimmed to only support the
Riva128 and perhaps some of the older chipsets.  The code has been tested on
several nVidia graphics card on x86 and x86_64.  I'm still waiting for
feedback from Guido Guenther regarding ppc.

This need not go to mainline immediately, as I would prefer the code to
receive rigorous testing in the mm tree.

Signed-off-by: Antonino Daplas <adaplas@pol.net>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>