[PATCH] ppc32: uninorth-agp suspend support
This patch adds suspend/resume support to the Apple UniNorth AGP bridge to
make sure AGP is properly disabled when the machine goes to sleep.  Without
this, the r300 based laptops will fail to wakeup from sleep when using the
new experimental r300 DRI driver.  It should also improve reliablility in
general with other chips.

Unfortunately, uninorth-agp is just a "sibling" of the video chip on the
PCI bus, and thus ends up beeing called either before the video chip
suspend routine, or after, depending on the HW layout or other random
things.

To make sure the device side of AGP is always disabled first and that we
never touch the device after having put it into D2 state (which can be
deadly), I also need the separate patch to radeonfb and aty128fb which will
make them disabled their own side if not already done by the bridge driver.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>