[PATCH] ppc64: don't use in_atomic()
in_atomic() in viopath.c was just used to determine if we had initialised
enough to be able to wait in a semaphore (i.e.  schedule).  Thus it can be
replaced now with checking system_state for SYSTEM_RUNNING.

Test booted on iSeries (which is the only place it is used).

Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>