[PATCH] ipr: Fast failure module options
Since the ipr adapter initialization is performed by waiting for
an interrupt to occur and this is the first interrupt ever received
from the adapter, and the timeout for this interrupt is 5 minutes,
failed ipr adapters can halt the boot process for a long time.
In certain environments, it makes sense to not try as hard to get
the adapter working and just fail the adapter. This patch adds
two module parameters to modify the adapter initialization sequence.

Signed-off-by: Brian King <brking@us.ibm.com>
Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>