[PATCH] USB: ehci and short in-bulk transfers with 20KB+ urbs
This changes handling of short bulk IN transfers with URB buffers that require
two or more transfer descriptors.  The case is rare in most systems, since few
drivers use such large buffers (bigger than 20KB, assuming 4K-aligned).

The existing code misbehave on at least NEC's EHCI 0.95 silicon, where it seems
to hit a new variant of a silicon quirk relating to dummy TDs.  Symptom of the
misbehavior is that the host stops polling on the IN endpoint, and the URB queue
no longer progresses.

This fix uses simpler logic to detect those short reads, dodging that quirk.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>