[IPV4]: Fix MTU check in ipmr_queue_xmit
Somehow I missed four files with dst_pmtu usages in them.  I'm going
to split them along the sames lines I did before: bug fixes and then
the trivial changes.

Here is a patch that replaces dst_pmtu with dst_pmtu in ipmr.c
since this is straight IPIP tunneling equivalent to what we have
in ipip.c.

As it is we may send ICMP packets when IPsec is present which is
exactly what the comment says that we shouldn't do.
 
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>