[ARM PATCH] 2575/1: pass -mbig-endian/-mlittle-endian to invocations of cpp
Patch from Lennert Buytenhek

We currently pass -mbig-endian/-mlittle-endian to the compiler and the
assembler, but we do not pass it to cpp, even though it also needs it.
So, add this flag to CPPFLAGS, and since CPPFLAGS is also passed to the
compiler and the assembler, we can then then remove the flag from CFLAGS
and AFLAGS.
The symptom of this problem is the kernel hanging after
"Memory: ......KB available (....K code, ...K data, ...K init)"
when compiling a big-endian kernel with a little-endian toolchain or
vice versa, and this happens because 'jiffies' points to the wrong
half of 'jiffies_64' (due to vmlinux.lds having been generated with
the wrong preprocessor flags), so calibrate_delay() takes forever to
complete.
Derived from an earlier patch by Krzysztof Halasa.

Signed-off-by: Lennert Buytenhek
Signed-off-by: Russell King