[IPV4]: Clear DF bit in ip_fragment fast path
It is possible for ip_fragment() to send out head fragments with
both DF and MF set for packets with local_df set to true.  This is
because the fast path tries to only modify the MF bit of the head
fragment.

Since the offset is always zero for the head fragment, and we know
that DF should be cleared in case of local_df, we can change |= to
a straight assignment.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>