[PATCH] USB: usbmon - document and kill pipe from API
Someone at linux-usb-devel observed that usbmon should not leak the concept
of pipe to the user space. This patch makes a compromise between having
too many words in the line and the pipe. It breaks API, but I doubt anyone
has any usbmon tools in circulation at this time, so I decided not to move
to a different filename.

Also, I decided to add a document which describes what is going on. I was
reluctant to do it because I doubt my ability to keep it synchronized and
updated, but there were requests to have one.

Signed-off-by: Pete Zaitcev <zaitcev@yahoo.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>