[PATCH] USB Storage: make usb-storage structures refcounted by SCSI
This patch started life as as474 from Alan Stern.  It's been rediffed
against the tip, tho that is now several days old.

This patch changes the way our private struct us_data is allocated; now it
gets stored at the end of the Scsi_Host rather than separately.  That's
what the hostdata field is intended for, and this is how other low-level
host drivers operate.  In order to convert between us_data and the
corresponding Scsi_Host I added two new inline routines: us_to_host and
host_to_us.  (The conversion actually should be quicker than before by a
microscopic amount, because now it only involves adding an offset whereas
before it involved dereferencing a pointer.)

The main advantage is that the host is refcounted, so now our us_data
automatically is too.  Although that doesn't matter at the moment, it will
matter later on when the control thread may need to outlive the disconnect
callback.


Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Matthew Dharm <mdharm-usb@one-eyed-alien.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>