[PATCH] x86: CMOS time update optimisation
This patch changes the update of the cmos clock to be timer driven rather
than poll driven by the timer interrupt function.  If the clock is not
being synced to an outside source the timer is removed and thus system
overhead is nill in that case.  The update frequency is still ~11 minutes
and missing the update window still causes a retry in 60 seconds.

We want the calls to sync_cmos_clock() to be made in a consistent environment.
This was not true when calling it directly from the NTP call code.  The
change means that sync_cmos_clock() is ALWAYS called from run_timers(), i.e.
as a timer call back function.

Also, call the timer code only through the timer interface (set a short timer
to do it from the ntp call).

Signed-off-by: George Anzinger <george@mvista.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>