[PATCH] ES7000 Legacy Mappings Update
This update only affects Unisys' ES7000 machines.

The patch reflects a change needed to determine which generation of ES7000
is currently running.  The next generation of ES7000s will have
conventional legacy support so the patch accommodates for this.  This patch
has been tested and verified on both an authentic 5xx ES7000 box and the
next generation ES7000 box.

Signed-off-by: Natalie Protasevich <natalie.protasevich@unisys.com>
Signed-off-by: Jason Davis <jason.davis@unisys.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>