[PATCH] x86_64: Fix NMI RTC access race
Fix NMI RTC access race

This patch fixes a race between the CMOS clock setting and the NMI code.  The
NMI code indiscriminatly sets index registers and values in the same place the
CMOS clock is set.  If you are setting the CMOS clock and an NMI occurs, Bad
values could be written to or read from the CMOS RAM, or the NMI operation
might not occur correctly.

Resetting the NMI is not required on x86_64 (in fact, it should not be done
according to the ICH5 documentation).  This patch simply removes the useless
code.

Signed-off-by: Corey Minyard <minyard@acm.org>
Signed-off-by: Andi Kleen <ak@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>