[PATCH] uml: fix "cond. expr. as lvalues" warning
Gcc 3.4.3 (and probably any 3.4) emits some "deprecation" warnings currently
about usages of CHOOSE_MODE, since the below syntax has been deprecated:

(a ? foo: bar) = foobar;
which often results from expansion of:
CHOOSE_MODE(foo, bar) = foobar;

So add an additional __CHOOSE_MODE syntax for users which need to get a
lvalue, which uses (a ? &foo : &bar) inside a function, casts the result to
the correct type and dereference the pointer, and use it where needed (i.e.
in <sysdep/ptrace.h>)

The patch builds and runs correctly (this has been tested on i386 only, not on
x86_64), and removes all the warnings.

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade@yahoo.it>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>