[PATCH] s390: signal stack bug
If a signal handler is set to use the signal stack (SA_ONSTACK), but the
signal stack is disabled, the signal frame should be written to the current
stack without stack switching.

The reason for the bug is get_sigframe() using on_sig_stack() instead of
sas_ss_flags(), which would be ok.

Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>