[PATCH] Can't unmount bad inode
This patch fixes a problem when a inode which is the root of a mount
becomes bad (make_bad_inode()).  In this case follow_link will return
-EIO, so the name resolution fails, and umount won't work.  The
solution is just to remove the follow_link method from bad_inode_ops.
Any filesystem operation (other than unmount) will still fail, since
every other method returns -EIO.

A test case for this is:

 1) export an smbfs on host A and mount the share on host B

 2) create directory X on A under the exported directory

 3) bind mount X to Y on B (Y need not be under the share)

 4) remove directory X, and create regular file X (same name) on A

 5) stat X on B, this will make X a bad inode (file type changed)

 6) umount Y

Without the patch applied, umount won't succeed, and a reboot is
necessary to get rid of the mount.

With the patch applied, umount will succeed.

The same is true for any filesystem which uses make_bad_inode() to
mark an existing inode bad (NFS, SMBFS, FUSE, etc...).

Signed-off-by: Miklos Szeredi <miklos@szeredi.hu>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>