[PATCH] io_remap_pfn_range: fix some callers for XEN
(from Keir:)
I have audited the drivers/ and sound/ directories. Most uses of
remap_pfn_range are okay, but there are a small handful that are
remapping device memory (mostly AGP and DRM drivers).

Of particular driver is the HPET driver, whose mmap function is broken even
for native (non-Xen) builds.  If nothing else, vmalloc_to_phys should be used
instead of __pa to convert an ioremapped virtual address to a valid physical
address.  The fix in this patch is to remember the original bus address as
probed at boot time and to pass this to io_remap_pfn_range.

Signed-off-by: Keir Fraser <keir@xensource.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>