[PATCH] I2C: lost arbitration detection for PCF8584
[PATCH] lost arbitration detection for PCF8584 algo driver

Patch against a slightly-dated linux-2.6 BK tree

This patch provides lost arbitration detection for the PCF8584
I2C algorithm driver.  The PCF8584 LAB bit is set whenever lost
arbitration is detected, so we check the bit in the wait_for_pin
function and if LAB is detected we return -EINTR.  The -EINTR
value bubbles-up all the way to the master_xfer API call so
callers may detect this condition explicitly.  LAB could be checked
more often, at the expense of code readability/maintainability.

Signed-off-by: Eric Brower <ebrower@gmail.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>