[IPSEC]: Make IPCOMP more resilient.
Since the IPCOMP header is left off when the payload is incompressible
or too small, we can also do the same thing when we encounter an error
during compression.

In other words, we can let outbound IPCOMP always succeed.  In the cases
where it would currently fail we simply skip the IPCOMP transform.  This
makes IPCOMP slightly more resilient when memory is low and simplifies
the code quite a bit.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Acked-by: James Morris <jmorris@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>